==> Scripts/01-download_Chrystia_data.R <==
#### Preamble ####
# Purpose: Download @cafreeland Tweets via voson.tcn
# Author: Inessa De Angelis
# Date: 4 October 2023
# Contact: inessa.deangelis@mail.utoronto.ca 
# License: MIT

#### Workspace setup ####
library(voson.tcn)
library(tidyverse)

#### Authentication ####
token <- tcn_token(bearer = "ADD SECRET CREDENTIALS")

#### Chrystia Freeland - key word "climate" and "environment", tweets from October 2022 - October 2023 ####
# Collect tweets (this series is from May 13, 2022) #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1657437032073838594",
    "1657468214794948614",
    "1657525605347254272",
    "1657531592166842370",
    "1657512967888334848",
    "1657469187621543936",
    "1657446526904422406",
    "1657470491265753088",
    "1657749900736741376",
    "1657465796946456578",
    "1657497389798678534",
    "1657465966324850688",
    "1657778287836561408",
    "1657723634495156224",
    "1657604548821925888",
    "1657476665755283460",
    "1657498630662639617",
    "1657505534507974658",
    "1657509419805016064",
    "1657523963818278914",
    "1657523993640062976",
    "1657551179977719809",
    "1657551552679124992",
    "1657586705455812608",
    "1657752978022088707",
    "1657478944751222788",
    "1657558466217615360",
    "1657595525808656385",
    "1657458651722858497",
    "1657563327428710401",
    "1657557781178912768",
    "1657513976912044033",
    "1657517592863449095",
    "1657474136745168899",
    "1657487143114096643",
    "1657484485594578948",
    "1657766999047434242",
    "1657734986684923906",
    "1658121436085600256",
    "1657535909548490754",
    "1657848152236650496",
    "1657468152035561472",
    "1657533665461542914",
    "1657457384514519041",
    "1657800370025037824",
    "1657726347014201345",
    "1657564690669531136",
    "1657546628062904320",
    "1657602235507920897",
    "1657494782095048705",
    "1657474629081116672",
    "1657456032002584578",
    "1657451899560292354",
    "1657844084382527489",
    "1657785789911117826",
    "1657553855901495298",
    "1657486337933721600",
    "1657629845285916672",
    "1657444663526313984",
    "1657442425324060674",
    "1657576710521384960",
    "1657569855644762112",
    "1657822009059934208",
    "1657516665314201600",
    "1657514561854124032",
    "1657458577156734976",
    "1657765304645349376",
    "1657810299389345792",
    "1657462938905595905",
    "1657438189026672640",
    "1657750788436029440",
    "1657706597794062336",
    "1657595046668156928",
    "1657706924974956544",
    "1657495342332489729",
    "1657514383969230849",
    "1657510730764001285",
    "1663759403357552643",
    "1657555697670983681",
    "1657509447047208961",
    "1657723268223123456",
    "1658546986398908416",
    "1658803031696363520",
    "1657544142367633408",
    "1657514470619377664",
    "1657457739818311682",
    "1657491392745345025",
    "1657440949495005187",
    "1657755567102803969",
    "1657755657766838273",
    "1657514266516176898",
    "1657511614231334913",
    "1658194211500335108",
    "1658961845611376645",
    "1660189787469299712",
    "1657448848481210369",
    "1657599400603193345",
    "1657474257763332103",
    "1657477850520223747",
    "1658682353689194496",
    "1657458525269004289",
    "1657472891989168128",
    "1657475087849885696",
    "1657605904806674432",
    "1657475087849885696",
    "1657605904806674432",
    "1657869790692646918",
    "1657438239047786499",
    "1657438661862912002",
    "1657444206888058880",
    "1657457516937084928",
    "1659421298458755072",
    "1658559675988672512",
    "1657535498187857920",
    "1657483467611881481",
    "1657469312888717314",
    "1657586469333024768",
    "1657462120802287616",
    "1658921023763349509",
    "1657488420384522242",
    "1657466854665863169",
    "1657465880584900609",
    "1657465640880414721",
    "1657481328722411520",
    "1657974665720238085",
    "1657506851792076802",
    "1657764868626485248",
    "1657472425578119174",
    "1657439598631346176",
    "1657476995914035202",
    "1657776374285737990",
    "1657447586213576705",
    "1657711002266267648",
    "1657541033331236866",
    "1657729315738402817",
    "1657744009383624706",
    "1657464612734791681",
    "1657584152043036673",
    "1657718843454681089",
    "1657476158999482379",
    "1657458441735028738",
    "1657551167923318784",
    "1657534969747591170",
    "1657535223750336514",
    "1657539640503443457",
    "1657492136911073283",
    "1657701129965248514",
    "1657529994719944704",
    "1657542042757496836",
    "1657537543930257410",
    "1657535929081380864",
    "1657738563092619264",
    "1657500679408066560",
    "1657714988146982914",
    "1657727721437556736",
    "1657735696864665601",
    "1657756132780265478",
    "1657739147694538752",
    "1657477671922855936",
    "1657549887209365505",
    "1657549887209365505",
    "1657552407776075778",
    "1657716589687459841",
    "1657469830738464772",
    "1657542805164806144",
    "1657463342615724032",
    "1657530914585575425",
    "1657490530027143178",
    "1657683444963463168",
    "1657495107262722048",
    "1657543541831376896",
    "1657475024373071876",
    "1657770824223301636",
    "1657483959909990403",
    "1657770262849536005",
    "1657477145994600448",
    "1657569181368238082",
    "1657468966980186114",
    "1657593875706568707",
    "1657486009762816001",
    "1657479946665857024",
    "1657477924356849665",
    "1657506066693869568",
    "1657456958935293953",
    "1657438522012467201",
    "1657443882941190144",
    "1657449131919515648",
    "1657499633105670144",
    "1657893910205849600",
    "1657920850425774081",
    "1657441269604311047",
    "1657437396722413570"
  )

thread_tweets <- tcn_threads(tweet_ids, token)

print(thread_tweets)

as_tibble(thread_tweets[["tweets"]])

as_tibble(thread_tweets[["users"]])

write_csv(x = thread_tweets[["users"]],
          file = "raw_thread_tweets3.csv")

write_csv(x = thread_tweets[["tweets"]],
          file = "raw_thread_tweets4.csv")


# Collect tweets #
# This might have the worst content of all replies? #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1656806355053735936",
    "1657120563670798349",
    "1656826328539365376",
    "1656869997531238405",
    "1656827968306413570",
    "1656900517652570113",
    "1657188772826558465",
    "1657052867986071552",
    "1660527941355204613",
    "1657538677264859136",
    "1656822053742182401",
    "1656821910548652035",
    "1657105826283417600",
    "1657028991776063492",
    "1656891858398908416",
    "1657170185806135302",
    "1657164638276993024",
    "1656817978355097604",
    "1656816984456155139",
    "1656817255722758144",
    "1657056241200545793",
    "1656928208355459072",
    "1656885920635793408",
    "1656842679689400322",
    "1657141527293087744",
    "1656926630982418433",
    # This is a level 7 #
    "1657130045775790080",
    "1657014369350287363",
    "1656809956937707521",
    "1656807805788061696",
    "1656809611926679554",
    "1656815338783387649",
    "1657102757948538881",
    "1657015772592455683",
    "1657178678554423296",
    "1657136924497178627",
    "1657094542984871936",
    "1656845056098041857",
    "1656966385149308929",
    "1656870792955559937",
    "1656869090735079425",
    "1657125200918458368",
    "1657010079080542208",
    "1656812473100451841",
    "1657006981130125313",
    # manipulated media #
    "1657009096149565440",
    "1656841898965848065",
    "1656810271158198273",
    "1656868573925756928",
    "1657492793252777984",
    "1657147745919008772",
    "1657407567340113921",
    "1657043767415783424",
    "1657110801281613824",
    "1656828127971246081",
    "1657012153868599299",
    "1656997536010317824",
    "1656808941752586242",
    "1656873945231134720",
    "1656807704214622209",
    "1656807776897703936",
    "1656808039674966016",
    "1656982271088271360",
    "1657132348385939456",
    "1656826684157603842",
    "1656825374544900096",
    "1656997407857442816",
    "1657103572234903552",
    "1657000144233984002",
    "1656808721115234306",
    "1656816583262498817",
    "1657069732229713920",
    "1656831976064647168",
    "1656832570305249281",
    "1656860344025862146",
    "1656972577863081986",
    "1657033759034683393",
    "1657033801292283905",
    "1657113181033222144",
    "1656843057873010689",
    "1656840453809995776",
    "1657126768233308162",
    "1656838137216131072",
    "1656952622509867010",
    "1657030732231983108",
    "1656813249759698952",
    "1657029465522700290",
    "1657016039769772032",
    "1656996159574114305",
    "1656983596769460224",
    "1656978426857652224",
    "1657172173373489153",
    "1657035036544102401",
    "1657090883563446281",
    "1657132627256565761",
    "1656861806659502081",
    "1656820494811709440",
    "1656848680400592897",
    "1656819511977459714",
    "1656831717725048833",
    "1656807102877437952",
    "1657097062192910337",
    "1657127849428500480",
    "1657128465156521984",
    "1656987360066772992",
    "1657031665556811781",
    "1657000030165934080",
    "1656821339762688001",
    "1657141764522921984",
    "1656960668174622720",
    "1657010803214635010",
    "1657426118939484160",
    "1657006329872125954",
    "1657071858506661908",
    "1656809849811013632",
    "1657513350111313920",
    "1656972989542400003",
    "1656813504106463237",
    "1656811018465705985",
    "1656890372126285827",
    "1657135207747817472",
    "1656832992785076224",
    "1656840883700916225",
    "1656883471279878150",
    "1656904762216923141",
    "1657008026673455105",
    "1656806629134811139",
    "1657114443191406598",
    "1656826114675818502",
    "1656826249350660101",
    "1656826351091941377",
    "1656826385426526208",
    "1657012835220586496",
    "1656983828374716419",
    "1656862544274026496",
    "1656952313716817920",
    "1656881180044677120",
    "1656881283962748930",
    "1656881342259376128",
    "1656881396663697409",
    "1656881687853236227",
    "1656882610574950401",
    "1656882672105299971",
    "1656882764392587264",
    "1656882836643692544",
    "1656883642654765057",
    "1656807057771880448",
    "1656994089991110658",
    "1657165906869452801",
    "1657026815104364544",
    "1656868848900145152",
    "1656994615285673985"
  )

chrystia_tweets_may_11 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_may_11)

as_tibble(chrystia_tweets_may_11[["tweets"]])

as_tibble(chrystia_tweets_may_11[["users"]])

write_csv(x = chrystia_tweets_may_11[["users"]],
          file = "raw_chrystia_tweets_may_11.csv")

write_csv(x = chrystia_tweets_may_11[["tweets"]],
          file = "raw_chrystia_tweets_may_11_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1656692429561421826",
    "1657025553327616001",
    "1656699258278670336",
    "1656704608729956352",
    "1656695658923937792",
    "1656734597844197378",
    "1656734597844197378",
    "1656772198474870787",
    "1656693521129021440",
    "1656719456138313728",
    "1656755297762050048",
    "1657106788087640065",
    "1656697369726009345",
    "1656720939747246089",
    "1656720735115804672",
    "1656848729129775105",
    "1656693068144295937",
    "1656721200322842626",
    "1656751496367861760",
    "1656813566484201474",
    "1656693268082556929",
    "1656789695156699138",
    "1656735932966371349",
    "1656692734319824896",
    "1656702066260160514",
    "1656695348771840001",
    "1656725120344543238",
    "1656716459588845591",
    "1656698831101214736",
    "1656812256536170496",
    "1657280389981806592",
    "1656970086027251716",
    "1657376675716378625",
    "1656953396216180736",
    "1656886946155880448",
    "1656694849075048449",
    "1656828086044999681",
    "1656693588976103424",
    "1656761035007619078",
    "1656761270865915912",
    "1656790569736368128",
    "1657049762145931265",
    "1657586226311081984",
    "1656838219600740356",
    "1656738513713233928",
    "1656725258953474049",
    "1656811168546144256",
    "1656811205057445888",
    "1656811245113147392",
    "1656694334106882048",
    "1656712316145238016",
    "1656692544338554880",
    "1656746115008729089",
    "1656701264560881664",
    "1656885736065449984",
    "1656704147197136897",
    "1657016153481555974",
    "1656714885898838017",
    "1656717247065346048",
    "1656747666020372480",
    "1656703435213877248",
    "1657402375680712705",
    "1656770488805326855",
    "1656729230376304648",
    "1656736236109987840",
    "1656719295408140288",
    "1656759682386522120",
    "1656863214184046592",
    "1656875388285636609",
    "1656705341923508225",
    "1656741668459270144",
    "1657354702340890627",
    "1657361968527753216",
    "1657195262333407240",
    "1656750100310290446",
    "1656987849365880836",
    "1657244992685867010",
    "1657352783513763840",
    "1657352783513763840",
    "1657384543005843457",
    "1657384981394489345",
    "1657388463212068864",
    "1657398509140844544",
    "1656704098614296576",
    "1656819205142978562",
    "1656864238500188163",
    "1656877950266925056",
    "1656876304451641344",
    "1656811418006568962",
    "1656693805557465089",
    "1656734315894407169",
    "1656817730962464773",
    "1656731026666946560",
    "1657493702296866818",
    "1657930613091966976",
    "1656724148087926790",
    "1656740475716399109",
    "1656708357200416770",
    "1656990823609761792",
    "1656787868516220985",
    "1657134935088705536",
    "1656797889690058757",
    "1656977415124426752",
    "1656707894249766918",
    "1656740457844719616",
    "1656697923906813963",
    "1656714446310342675",
    # this is a level 6 #
    "1656711427619536919",
    "1656720927953145856",
    "1656730163554418713",
    "1656736908968611845",
    "1656739908508352512",
    "1656742119833313280",
    "1656819311762087937",
    "1656824860885000192",
    "1656832249533284354",
    "1656834148512968704",
    "1656853090039898113",
    "1656874503606091778",
    "1656875740678217729",
    "1656732026689716229",
    "1656873580716519424",
    "1656729657704841216",
    "1656879323109003264",
    "1657052584325140483",
    "1657183520203108353",
    "1656704688274931712",
    "1657016285736341504",
    "1656705964064514048",
    "1656706332748070912",
    "1656789194902224897",
    "1656872568106946560",
    "1656725272308137985",
    "1656694261268758530",
    "1656756407470006273",
    "1656952049697693698",
    "1656712110024409100",
    "1656734248785543181",
    "1656743475436736512",
    "1656718481948282882",
    "1656745098208501760",
    "1656743771617259521",
    "1656745500245389312",
    "1656737720872624129",
    "1656693414740541441",
    "1656811731254124544",
    "1656695715790192641",
    "1656722280528117761",
    "1656767150668001283",
    "1656881020258467840",
    "1657025441096507397",
    "1658498722043494405",
    "1656734155437387776",
    "1656846602386939904",
    "1656826678403031040",
    "1656827088719196160",
    "1656711370736386060",
    "1656972442235985920",
    "1656842580234219521",
    "1657021520311382027",
    "1656697103614197760",
    "1656784390867898369",
    "1657191637364211715",
    "1657040861950033928",
    "1656716271243624464",
    "1656702323425681415",
    "1656693710950891521",
    "1656697646478946304",
    "1656712747931910153",
    "1656855052689047552",
    "1656719125710790656",
    "1656731496013545473",
    "1656852475276673024",
    "1656709020462247936",
    "1656734104941887500",
    "1656876039371657216",
    "1656969542097043457",
    "1656987428480249856",
    "1656698381207470080",
    "1656743744337739777",
    "1656701341287301123",
    "1657096224640491524",
    "1656693037739671555",
    "1656693092932632576",
    "1656692708667191298",
    "1656699769962545152",
    "1656699794063081480",
    "1656699853764739072",
    "1656699925361573894",
    "1656699986791350281",
    "1656700105016197125",
    "1656700146518720512",
    "1656700216127504396",
    "1656700269655212040",
    "1656692849092489216",
    "1657023195608645634",
    "1656743733029748736",
    "1656725644422811649",
    "1656877743349563393",
    "1656750848385359872",
    # this is a level 5 #
    "1656695077971054593",
    "1656815871996878848",
    "1656896755106209792",
    "1656995455899779073"
  )

chrystia_tweets_may_11_2 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_may_11_2)

as_tibble(chrystia_tweets_may_11_2[["tweets"]])

as_tibble(chrystia_tweets_may_11_2[["users"]])

write_csv(x = chrystia_tweets_may_11_2[["users"]],
          file = "raw_chrystia_tweets_may_11_3.csv")

write_csv(x = chrystia_tweets_may_11_2[["tweets"]],
          file = "raw_chrystia_tweets_may_11_4.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1620255649342029825",
    "1620286651175141377",
    "1620282239375253504",
    "1620497925532000279",
    "1620270709628112897",
    "1620435438350569472",
    "1620304564108025856",
    "1620257865989763072",
    "1620380206438707201",
    "1620301663293825024",
    "1620307575890919426",
    "1620265978558054400",
    "1620443142221602818",
    "1622679648248246273",
    "1620476161087401986",
    "1620580395325755392",
    "1620478292162928640",
    "1620371164924379141",
    "1620286110403551234",
    "1620655030771531776",
    "1620282380870377472",
    "1620619931036311552",
    "1620293798504054785",
    "1620297621272252417",
    "1620538635820109824",
    "1620259512031793153",
    "1620260581554135040",
    "1620277048366489602",
    "1620278189724680193",
    "1620305612427993088",
    "1620258583136075779",
    "1620282505537675265",
    "1620515450940538894",
    "1620361016906493953",
    "1620565608025980929",
    "1620510151466942464",
    "1620266163577188352",
    "1620277042276335616",
    "1620261076637224963",
    "1620497959443128323",
    "1620297495095173120",
    "1620259631833702402",
    "1620721415661187072",
    "1620257691204751363",
    "1620276540205572096",
    "1620257948730630144",
    "1624148199710818304",
    "1620616176148811777",
    "1620607391023906817",
    "1620581851331977217",
    "1620275126637064192",
    "1620267515023851520",
    "1620256862561247232",
    "1620530997782806529",
    "1620573848226897920",
    "1620792073170591744",
    "1620751329722523648",
    "1620260721593577474",
    "1620416914337169408",
    "1620431267249655808",
    "1620277495751921664",
    "1620318502153498629",
    "1620418000972308480",
    "1620454328040583168",
    "1620336199310405632",
    "1620451263405232128",
    "1620380066327965697",
    "1620411610807537667",
    "1620259074804961281",
    "1621039179986788352",
    "1621174939976859650",
    "1620259597163589632",
    "1620261546906746882",
    "1620455922605912064",
    "1620449610904014848",
    "1620374471940124672",
    "1620288275255808001",
    "1620288438451974145",
    "1620328776759640064",
    "1620464782808879105",
    "1620286619709485059",
    "1620440444734967808",
    "1620291151155838976",
    "1620377722122813441",
    "1620378570185916419",
    "1620405817622102017",
    "1620445660456833027",
    "1620441695266689024",
    "1620374548179992577",
    "1620285096854183936",
    "1620434138712133632",
    "1620467997478621184",
    "1620290702398881792",
    "1620358358145572864",
    "1620288165327273986",
    "1620682376534949889",
    "1620263990516666368",
    "1620264049098493953",
    "1620264108787654659",
    "1620264162529247233",
    "1620264229269041153",
    "1620296612336791552",
    "1620335336038764544",
    "1620438587899789313",
    "1620463386965471232",
    "1620600829328887808",
    "1620601876931645441",
    "1620845443126296577",
    "1620334143170289665",
    "1620290689740464128",
    "1620265945158799360",
    "1620595528294871040",
    "1620299680268836865",
    "1620373209421074432",
    "1620313213694021633",
    "1620591755220467712",
    "1620442034938200067",
    "1620347018748235777",
    "1620990513498161152",
    "1620394660626862080",
    "1620461812595695617",
    "1620566578956996608",
    "1620567342114824192",
    "1620326712960770050",
    "1620261615592685568",
    "1620284537942024192",
    "1620261045230239745",
    "1620261345521459200",
    "1620385412173955073",
    "1620454828307775488",
    "1620278007746412545",
    "1620613654617157633",
    "1620278368590774272",
    "1620262454931644416",
    "1620258289505427458",
    "1620374157069553664",
    "1620481375991390208",
    "1620954036554940417",
    "1620323472407945219",
    "1620291704724291584",
    "1620281414876012549",
    "1620574439808303104",
    "1620455562810101761",
    "1620278741078245376",
    "1620258771670044672",
    "1620259106014756870",
    "1620313208912482304",
    "1620306436764082178",
    "1620280982489436160",
    "1620592855952805888",
    "1620258211277438977",
    "1620584786623598592",
    "1621006640375029761",
    "1621006703532867586",
    "1621006768406151168",
    "1621006822315536386",
    "1621006876996694016",
    "1621006937839276032",
    "1621006996211388417",
    "1621007048875085827",
    "1621007102130126849",
    "1621007162029010944",
    "1620556360072204289",
    "1620265138782867456",
    "1620497970067300352",
    "1620282593127325696",
    "1620418978219786240",
    "1620441032159797250",
    "1620475137169707009",
    "1620579729870028801",
    "1620435187300700160",
    "1621007220832964608",
    "1621007282388762626",
    "1621007326432956416",
    "1620263261227876354",
    "1620482390580924416",
    "1621100613554827264",
    "1621101236048257024",
    "1621102580352057344",
    "1621111951232483329",
    "1620617535828918273",
    "1620353933976932354",
    "1620285553790054402",
    "1620414215273222147",
    "1620483479682322432",
    "1620258940494958592",
    "1620586842684030977",
    "1620742092426465282",
    "1620742263822508035",
    "1620573332050677761",
    "1620463093754236931",
    # brainless = level 3 #
    "1620589816072273921",
    "1620398125583327232",
    "1620385309333622785",
    "1620275124032405505",
    "1620499993823182849",
    "1620585979257171969",
    "1621662379145629703",
    "1622755112660742146",
    "1628584923610185734",
    "1632918775988211712"
  )

chrystia_tweets_jan_30 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_jan_30)

as_tibble(chrystia_tweets_jan_30[["tweets"]])

as_tibble(chrystia_tweets_jan_30[["users"]])

write_csv(x = chrystia_tweets_jan_30[["users"]],
          file = "raw_chrystia_tweets_jan_30.csv")

write_csv(x = chrystia_tweets_jan_30[["tweets"]],
          file = "raw_chrystia_tweets_jan_30_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1629547522254405632",
    "1629736027714732035",
    "1629606535465959426",
    "1629578580987629568",
    "1629914001663733760",
    "1629564360225099776",
    "1629560002208727040",
    "1629628003541024770",
    "1629709554672959491",
    "1630748612077694981",
    "1629548049054801921",
    "1629588077533339648",
    "1629608016126328832",
    "1629583385801764871",
    "1629639684216303617",
    "1629678024403689474",
    "1629677313821638657",
    "1630249780755701761",
    "1629586055224590339",
    "1629655521274785792",
    "1629713431610941441",
    "1629910580353200128",
    "1629610647565606912",
    "1630247740860157957",
    "1629840361362280449",
    "1629998457028059136",
    "1629721107145150465",
    "1629707033367957505",
    "1629625855298134016",
    "1629670145504464896",
    "1629623452574396418",
    "1630086729964359680",
    "1629875667457425410",
    "1629675786092392451",
    "1629593449149382656",
    "1629547714064121858",
    "1629816968206270467",
    "1630050207923617794",
    "1630086255353663493",
    "1630144570871930880",
    "1629826928705249281",
    "1630724532058333184",
    "1629717550547312640",
    "1629664339186659328",
    "1629664339186659328",
    "1629639189271543810",
    "1630240782862127104",
    "1630113580027908101",
    "1629550916566151169",
    "1630565436050796545",
    "1629873341329727489",
    "1630030005475368960",
    "1630979885194936321",
    "1629764689608728576",
    "1629644588498468864",
    "1629622331621642240",
    "1629704871283773442",
    "1630235939166625792",
    "1630789480960409600",
    "1629561077896818688",
    "1629547709366452225",
    "1629640704036339713",
    "1630754783677108224",
    "1629551289234522112",
    "1629583083920826368",
    "1629583093257453570",
    "1629664940276498434",
    "1629571181480235009",
    "1630106971583164422",
    "1629703771461754880",
    "1629882157329190912",
    "1629692984366190592",
    "1629753900738674688",
    "1632073957552668672",
    "1629875087825584128",
    "1629714273843773442",
    "1629752433290223621",
    "1629647358400114688",
    "1629856390880854016",
    "1629649240107663361",
    "1629553956677472256",
    "1629548102204915712",
    "1630194141492641793",
    "1630038433111625730",
    "1629649132183863297",
    "1629697536272498689",
    "1629645791672270849",
    "1629663338207809536",
    "1629624536135024641",
    "1629608211975135235",
    "1629548084022714368",
    "1629632612632260611",
    "1629866924237807619",
    "1630971149252165637",
    "1631063060302495744",
    "1629612335810727939",
    "1648014210058559507",
    "1629657427422453760",
    "1629681880709693443",
    "1629623412011180034",
    "1629558426081673218",
    "1629617802507067400",
    "1629737720334680064",
    "1629771458296827907",
    "1629561522417532930",
    "1630407431355523072",
    "1629706562989088770",
    "1629697861935075329",
    "1629621657521336320",
    "1629551104223526914",
    "1629622149685051393",
    "1629706518290640896",
    "1629724311291559937",
    "1629817115602452483",
    "1629759950028410881",
    "1629615116260966406",
    "1629601338978971651",
    "1629569031505313792",
    "1629644248621494274",
    "1630411046057050112",
    "1630146460607217664",
    "1630720116127047682",
    "1631327932063916033",
    "1629884885975302145",
    "1630066317431562240",
    "1629950470692409344",
    "1630720038566232066",
    "1630717516057317376",
    "1629902012975546368",
    "1629844588532826112",
    "1629628256025452544",
    "1629629457899360257",
    "1629611105667489792",
    "1629641300428623872",
    "1629868531285381122",
    "1629783220676952064",
    "1629652748646989825",
    "1629593098757226496",
    "1629704193719123971",
    "1629778760135512064",
    "1629558927938703360",
    "1630725918053285889",
    "1629608340086222848",
    "1629625481929580547",
    "1629602030653263873",
    "1629866454513250305",
    "1629610525217751041",
    "1629741959442038784",
    "1629548738753470465",
    "1629692471595638785",
    "1629848850214027270",
    "1629678051876573184",
    "1629586060631224321",
    "1629874695804780551",
    "1629696593367560192",
    "1631164372117762048",
    "1629587242770665473",
    "1630714394031316994",
    "1629797088031830016",
    "1629554826047631360",
    "1630366037777477634",
    "1629925348082335744",
    "1630725359451926528",
    "1629886010849914880",
    "1629606262131380226",
    "1630715259723083776",
    "1629555918856761344",
    "1629582616692129795",
    "1629597134482993152",
    "1629707310565142529",
    "1629556050528735232",
    "1629692505641062400",
    "1629603442690777091",
    "1629734939481473025",
    "1630474477091065857",
    "1629600859960258562",
    "1629581177878110208",
    "1629622302768766976",
    "1630733576865218563",
    "1629865164882231300",
    "1629822551441514496",
    "1629855830471475201",
    "1630075370387980288",
    "1629686933516328962",
    "1629553070819385345",
    "1629876648773402626",
    "1629991246675542019",
    "1630721868612407296",
    "1630026605153206273",
    "1629604655931043840"
  )

chrystia_tweets_feb_25 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_feb_25)

as_tibble(chrystia_tweets_feb_25[["tweets"]])

as_tibble(chrystia_tweets_feb_25[["users"]])

write_csv(x = chrystia_tweets_feb_25[["users"]],
          file = "raw_chrystia_tweets_feb_25.csv")

write_csv(x = chrystia_tweets_feb_25[["tweets"]],
          file = "raw_chrystia_tweets_feb_25_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1616215053996285952",
    "1616266647547052033",
    "1616434178341363714",
    "1616451482852921348",
    "1616452058961281027",
    "1616491186398822400",
    "1616602042390450177",
    "1616368787762319360",
    "1616228976938352640",
    "1616323168674910210",
    "1616365991566491650",
    "1616404230599999490",
    "1616497098362429442",
    "1616551282541469696",
    "1616254600998440960",
    "1616454826988118019",
    "1616460269588471809",
    "1616576307411705856",
    "1616551731290148864",
    "1616497473253478410",
    "1616499344097816576",
    "1616249012558139392",
    "1616478088514408450",
    "1616218698980663296",
    "1616418754652504067",
    "1616258516964237313",
    "1616223077603368960",
    "1616232173832671232",
    "1616233796805840896",
    "1616242314535378949",
    "1616245760458383360",
    "1616501242569822209",
    "1616215425699790849",
    "1616215849203007488",
    "1616252451962912768",
    "1616253833721544705",
    "1616269207209136130",
    "1616274925274427393",
    "1616274979456172034",
    "1616278863150972930",
    "1616222805896597505",
    "1616312934732550144",
    "1616432785643933697",
    "1616270347904643077",
    "1616288018859724801",
    "1616299541158047744",
    "1616300443214258178",
    "1616301269831897090",
    "1616327969194246144",
    "1616329885265305601",
    "1616461666618675201",
    "1616407653089177603",
    "1616218915859845122",
    "1616233431863812097",
    "1616415669234405377",
    "1616416561891528704",
    "1616444361478393856",
    "1616449713993031680",
    "1616451994348302336",
    "1616468861527756801",
    "1616384309509619716",
    "1616221666199351297",
    "1616305935093497857",
    "1616321818918522882",
    "1616428827877511169",
    "1616431197621559298",
    "1616503644421062657",
    "1616528616422281217",
    "1616529250106093569",
    "1616535988628586514",
    "1616237922042126337",
    "1616379941603667968",
    "1616499604861685762",
    "1616258997715083264",
    "1616290756259348480",
    "1616415071730806786",
    "1616460342254616576",
    "1616502416718921731",
    "1616513441862193156",
    "1616242181991239682",
    "1616463773933633538",
    "1616467158191538181",
    "1616474224155316224",
    "1616487711262773251",
    "1616489498086342670",
    "1616500625520398344",
    "1616525589229305861",
    "1616532187154907141",
    "1616554616753213468",
    "1616630214406144000",
    "1616563981300432898",
    "1616295765877592064",
    "1616451862286467076",
    "1616459989106987009",
    "1616462815467896833",
    "1616517619149930500",
    "1616539468160462848",
    "1616608856326672385",
    "1616412292895367170",
    "1616560393773633538",
    "1616510953662873600",
    "1616285084599914500",
    "1616462728704720898",
    "1616517229151154177",
    "1616538500437868564",
    "1616218102303428608",
    "1616219124845117441",
    "1616220083398840325",
    "1616221617755140096",
    "1616225588087234561",
    "1616252457788805120",
    "1616253585272242177",
    "1616390086429401089",
    "1616401034364030976",
    "1616490281771192321",
    "1616486654809604098",
    "1616241581908185089",
    "1616260636031471617",
    "1616437319052976128",
    "1616499693319585792",
    "1616215595107897347",
    "1616235469561552896",
    "1616235647227908096",
    "1616248060429930496",
    "1616250411580813312",
    "1616262067421597697",
    "1616269030855266304",
    "1616417695926304770",
    "1616428901764567040",
    "1616433280580202496",
    "1616225060934524929",
    "1616227475712888832",
    "1616230554138968066",
    "1616231225580191744",
    "1616231712614146049",
    "1616235151809200128",
    "1616238377770323968",
    "1616242279479496706",
    "1616247642505318402",
    "1616248957788618752",
    "1616215445031505921",
    "1616216755763912706",
    "1616217528665079809",
    "1616230331274739713",
    "1616238551921737728",
    "1616241955486433283",
    "1616246994699423745",
    "1616250329577947136",
    "1616254035329421312",
    "1616260661129928706",
    "1616314523379699716",
    "1616515059965956098",
    "1616528968240463872",
    "1616289267738894338",
    "1616289346772152320",
    "1616289403571441666",
    "1616289445787103232",
    "1616289509720862720",
    "1616289638007836672",
    "1616289689325162496",
    "1616335535001448449",
    "1616368597227667459",
    "1616405909684785155",
    "1616427361712631810",
    "1616451038105698304",
    "1616427858813157381",
    "1616520415970734080",
    "1616270984926969858",
    "1616273928594522113",
    "1616400978865197058",
    "1616289575072321537",
    "1616289817960275970",
    "1616289991885455360",
    "1616470625773559814",
    "1616477628231188496",
    "1616365764193255426",
    "1616547527582699522",
    "1616499447776841728"
  )

chrystia_tweets_jan_19 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_jan_19)

as_tibble(chrystia_tweets_jan_19[["tweets"]])

as_tibble(chrystia_tweets_jan_19[["users"]])

write_csv(x = chrystia_tweets_jan_19[["users"]],
          file = "raw_chrystia_tweets_jan_19.csv")

write_csv(x = chrystia_tweets_jan_19[["tweets"]],
          file = "raw_chrystia_tweets_jan_19_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1616215053996285952",
    "1615865589502722048",
    "1616011905398628352",
    "1615929158378606594",
    "1616171725510823937",
    "1615872701930905601",
    "1615872810169307136",
    "1615872983905476608",
    "1615873087001575425",
    "1615873271328538626",
    "1615873384570576896",
    "1615873557493604352",
    "1615872573971070977",
    "1615873636241444865",
    "1615873791565123585",
    "1615873908896407552",
    "1615874045353902080",
    "1615968014192594947",
    "1616048857225256962",
    "1616233050668675072",
    "1615914443887476741",
    "1615972162988396544",
    "1616371679500070915",
    "1616457007581593601",
    "1616208652687314974",
    "1616097417061076998",
    "1616160561313775616",
    "1616173011182448640",
    "1615901514664394753",
    "1616266733416812549",
    "1615873782840758274",
    "1616039524349669377",
    "1616159912412725270",
    "1616195525048426496",
    "1615896749918519296",
    "1615866586182868992",
    "1616135337511731201",
    "1616121122503929857",
    "1616179017836544008",
    "1616098186493366274",
    "1616119860811165696",
    "1615880963506180096",
    "1615920554971901954",
    "1615964246768222211",
    "1615906287073591297",
    "1616176715386228737",
    "1616208185102315520",
    "1616163090634276864",
    "1615956951644659713",
    "1615899723730411520",
    "1615919846654353408",
    "1616195310329663489",
    "1615962835498172416",
    "1616190438934982656",
    "1616213664939270151",
    "1615934702917160962",
    "1615909661378584577",
    "1615866520826937344",
    "1615878368154554370",
    "1615891508099706880",
    "1615903420095016960",
    "1616077014502637569",
    "1616098852653727745",
    "1616205640988495872",
    "1615921339667460096",
    "1615879948035624963",
    "1615921847069212679",
    "1615986945926090752",
    "1615899036413837313",
    "1616223378922258433",
    "1615922731400429570",
    "1616103857549082625",
    "1616261584485027840",
    "1616261616403910656",
    "1615913669946380289",
    "1615866439600328704",
    "1616254420353941511",
    "1616223506777128960",
    "1616173368138661888",
    "1616151659184545792",
    "1616127473460350977",
    "1615896206689062915",
    "1616110097188651008",
    "1616175558232276992",
    "1616158796392300554",
    "1616178485516451848",
    "1615928528494542848",
    "1615866887967211522",
    "1615929757245349888",
    "1615870093950586880",
    "1615951959923752967",
    "1615904656609415169",
    "1615889823406899201",
    "1616183201889767424",
    "1616550983265550338",
    "1615922166222180353",
    "1615959274081128448",
    "1615941116590972929",
    "1616102217572024320",
    "1615867238535380993",
    "1615941567440723969",
    "1615956344456069120",
    "1616218697961558016",
    "1615920654733438977",
    "1615921444835454976",
    "1615920668058570754",
    "1616155978550083584",
    "1616139782899236877",
    "1615956513243410432",
    "1616173095584251914",
    "1616102504298868736",
    "1616035107118219268",
    "1616118302518161410",
    "1616038340519763971",
    "1616083445339365377",
    "1616154560812449797",
    "1616189053745106944",
    "1615948903353835520",
    "1616109477337468929",
    "1616174469025775627",
    "1615938395833831425",
    "1616059975649968130",
    "1615934875806363648",
    "1616167977262678016",
    "1616164695525974016",
    "1615912702014095363",
    "1616811032990650373",
    "1615942513231269894",
    "1616857135085932544",
    "1615868148313624578",
    "1615924663049224193",
    "1615919038198173703",
    "1615908895612108801",
    "1616224677600587776",
    "1616224909029511173",
    "1615922136824320001",
    "1616099791322361858",
    "1616124278889664518",
    "1616218387113476097",
    "1616180187410096128",
    "1616180275222020097",
    "1616102695525548033",
    "1616211163980926976",
    "1615871795579469830",
    "1616095241139851271",
    "1616206983094497281",
    "1616044651689185281",
    "1616044824683024386",
    "1616044991461339138",
    "1617256422311399426",
    "1616138167358201856",
    "1616095413035044865",
    "1615868063781425153",
    "1615868554485633028",
    "1616135195542814739",
    "1615877944706060289",
    "1616162148148715520",
    "1615883197488984065",
    "1615868510315454465",
    "1615929672587694080",
    "1616192987285049346",
    "1615915785536380928",
    "1616180143659094016",
    "1616108867016065026",
    "1615879509235978242",
    "1616141183570567168",
    "1616484477852606464",
    "1615955877877649410",
    "1615943161138008070",
    "1616122676065583116",
    "1616082311719833600",
    "1616179253636370432",
    "1616081152661413892",
    "1615956482461155328",
    "1616175620165091332",
    "1616175998201991168",
    "1616176277618049030",
    "1616176695341367324",
    "1616220664972865536",
    "1616093941811630082",
    "1616573285084307456",
    "1616172467604828161",
    "1616443680390549510",
    "1616094660090548225",
    "1619133098519846912",
    "1616207404630216763"
  )

chrystia_tweets_jan_18 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_jan_18)

as_tibble(chrystia_tweets_jan_18[["tweets"]])

as_tibble(chrystia_tweets_jan_18[["users"]])

write_csv(x = chrystia_tweets_jan_18[["users"]],
          file = "raw_chrystia_tweets_jan_18.csv")

write_csv(x = chrystia_tweets_jan_18[["tweets"]],
          file = "raw_chrystia_tweets_jan_18_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1580988645406027776",
    "1581004148501512194",
    "1581013257472864256",
    "1581081032291667968",
    "1581083706248220672",
    "1581090639789576193",
    "1581102038938972160",
    "1580996255186325504",
    "1581043649428660224",
    "1581088655300448256",
    "1581022072025014274",
    "1581090800821104640",
    "1581452491103014912",
    "1581148468193918978",
    "1580990598017785856",
    "1581138043943022592",
    "1581264799799087104",
    "1581348057404551169",
    "1580989676122673152",
    "1581110203130712064",
    "1581002211857162240",
    "1582010855545503744",
    "1580993396642304000",
    "1580990179657265152",
    "1580994586326941697",
    "1581339326608478210",
    "1580994529250795522",
    "1580996444055810048",
    "1581002839278526464",
    "1580991496198623234",
    "1580992181082300416",
    "1581004533039214592",
    "1581005561247989760",
    "1581024138755731456",
    "1581294532385472512",
    "1580989035530846209",
    "1580990767090511872",
    "1580994668350754816",
    "1580995743703519232",
    "1580996848382533632",
    "1580997592070361088",
    "1580989295958970368",
    "1580989630580944896",
    "1580993882623709189",
    "1580989505477423104",
    "1580991096846700545",
    "1580995830210650113",
    "1580998615564111872",
    "1580999091810865152",
    "1580999478500544513",
    "1581000210129354752",
    "1580990873986158593",
    "1580990961961750528",
    "1580998649634357249",
    "1580999782637907969",
    "1581000286763909121",
    "1581001062386786304",
    "1581004134341545984",
    "1581004553457057793",
    "1581004633186918400",
    "1581004682981363712",
    "1580992815344910336",
    "1580999187872636928",
    "1580999818666971136",
    "1581002041572589568",
    "1581004728737333249",
    "1581006271318458374",
    "1581007153221206016",
    "1581010464959782913",
    "1581011535987904513",
    "1580991785727569922",
    "1581010581318160384",
    "1581016847901945857",
    "1581021088074846208",
    "1581022818044502016",
    "1581024687270031361",
    "1581027439555538945",
    "1581028036564033538",
    "1581035448620974080",
    "1581039833669328896",
    "1581010185216491520",
    "1581011029927948290",
    "1581011037784272896",
    "1581014914751401985",
    "1581016755014897665",
    "1581017065942528000",
    "1581017973829300225",
    "1581021767224954881",
    "1581024416347353088",
    "1581032420924878848",
    "1580996659311280131",
    "1581003986223927296",
    "1581004830193352705",
    "1581008022746587136",
    "1581012487935520769",
    "1581016672252854272",
    "1581021390815100928",
    "1581034783584399360",
    "1581041627829002242",
    "1581043249136824320",
    "1581006212992466952",
    "1581009921247956993",
    "1581023504430166016",
    "1581029118157942784",
    "1581033032508547073",
    "1581038031993462784",
    "1581039584460230656",
    "1581046475663908864",
    "1581047706239504384",
    "1581059216365604865",
    "1581096311012880384",
    "1581096433255534593",
    "1581102910381125634",
    "1581001648083980288",
    "1581005038809645068",
    "1581005096254550016",
    "1581011017701920769",
    "1581013661552107520",
    "1581013991060811776",
    "1581390295089807360",
    "1580993877087248384",
    "1580995373975633920",
    "1581012534626496517",
    "1581014031489716224",
    "1581014304899223554",
    "1581014483824017408",
    "1581014855322021888",
    "1581015252615172096",
    "1581016143502127106",
    "1581005287603175429",
    "1581005629904543744",
    "1581028574676815872",
    "1581040533044367360",
    "1581041973544517632",
    "1581045040939663360",
    "1581045927561035778",
    "1581047643538853889",
    "1581063774945304576",
    "1581063807194890240",
    "1581014427171975168",
    "1581015665812860928",
    "1581017122938900481",
    "1581017899971776512",
    "1581018775578501120",
    "1581019539407798272",
    "1581020770607988736",
    "1581023314025119744",
    "1581023652270964736",
    "1581026050779512832",
    "1581335349019582464",
    "1581030016590434304",
    "1581059581949509632",
    "1581061208668401664",
    "1581064596223569920",
    "1581067282834288640",
    "1581069905956581376",
    "1581071078897897472",
    "1581074031381721088",
    "1581076004696903680",
    "1580995921059643392",
    "1581009342958284800",
    "1581037386209046528",
    "1581075609673170944",
    "1581084929752190976",
    "1581263780448595968",
    "1580991431304695808",
    "1580996687203827712",
    "1581034266653528064",
    "1581102862469566464",
    "1581022850412347394",
    "1581053985553022977",
    "1581062873207701504",
    "1581065316079382529",
    "1581066649251504128",
    "1581077825666547712",
    "1581080189328453632",
    "1581084443653337090",
    "1581085104700411904",
    "1581086823669190656",
    "1581004835943424001",
    "1581013555150983168",
    "1581018773116493824",
    "1581025684176375808",
    "1581028108878376960",
    "1581031323921420289",
    "1581032404676145152",
    "1581032520069840896",
    "1581033264789061632"
  )

chrystia_tweets_oct_14 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_oct_14)

as_tibble(chrystia_tweets_oct_14[["tweets"]])

as_tibble(chrystia_tweets_oct_14[["users"]])

write_csv(x = chrystia_tweets_oct_14[["users"]],
          file = "raw_chrystia_tweets_oct_14.csv")

write_csv(x = chrystia_tweets_oct_14[["tweets"]],
          file = "raw_chrystia_tweets_oct_14_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1580335942581514240",
    "1580583093554946048",
    "1580458394929876992",
    "1580645842389585920",
    "1580551332725133313",
    "1581271634358730752",
    "1580590422232924161",
    "1580343099867672577",
    "1580470386801537024",
    "1580391665357180929",
    "1580986770992156674",
    "1580689746514575360",
    "1580366405165539330",
    "1580620128609132544",
    "1580419093492289536",
    "1580678876950278144",
    "1580631890574606336",
    "1580706127574368259",
    "1581407775758110720",
    "1580586945935478784",
    "1580336163054792704",
    "1580339786770153473",
    "1580347727312916482",
    "1580366039145410560",
    "1580398388134449153",
    "1580586656037756934",
    "1580361893579214848",
    "1580338496211800064",
    "1580352361243226112",
    "1580363302802427905",
    "1580801461071400961",
    "1580649361788538880",
    "1580599796481789953",
    "1580959816792408073",
    "1580535004664590338",
    "1581041216309035008",
    "1580708239624523776",
    "1580625209790697472",
    "1580399589634371584",
    "1580371720870932480",
    "1580386239961710592",
    "1580589408960794624",
    "1580579028838150145",
    "1580356732173574145",
    "1580397350396506112",
    "1580336237457854464",
    "1580376736352518144",
    "1580343472086986752",
    "1580622163706806273",
    "1580408210225496064",
    "1580440035262099456",
    "1580696717305204740",
    "1580577941955571714",
    "1580701799484633088",
    "1580438173460881408",
    "1580731415964286976",
    "1580526202661396482",
    "1580405410133311488",
    "1580345312610512896",
    "1580867039476211714",
    "1580338724981702656",
    # Level 1 #
    "1580352670636052480",
    "1580550654640037889",
    "1580409674012438529",
    "1580403159184261120",
    "1580353836161474560",
    "1580383925620609024",
    "1580336172299341824",
    "1580646926990860288",
    "1580349889447309312",
    "1580358351283978240",
    "1580533770817777664",
    "1580679967599661056",
    "1580688163756212225",
    "1580648347374735360",
    "1580463387569704960",
    "1580420120207581186",
    "1580581154826387456",
    "1580401192504819712",
    "1580691397036101632",
    "1580762755120103426",
    "1580596832182206468",
    "1580417249042608130",
    "1581372525262479363",
    "1580656563533987840",
    "1580656639379591168",
    "1580661795814973440",
    "1580524382174466048",
    "1581000854949490689",
    "1580620904018907136",
    "1580659235628937216",
    "1580635554676637698",
    "1580621556509999105",
    "1580608839216660480",
    "1580636374633283584",
    "1580588721941815297",
    "1580526516638605313",
    "1580539526883467265",
    "1580425657125044224",
    "1580580593280389122",
    "1580668621096828928",
    "1580519365748740096",
    "1580499178253586433",
    "1580344007028510721",
    "1581103247770931200",
    "1580664168122048512",
    "1580689364480188416",
    "1580643879816429568",
    "1580646289028837376",
    "1580585342918623232",
    "1580703408171778050",
    "1580380898994032640",
    "1580546387124449280",
    "1580498088627929089",
    "1580621864384098306",
    "1580703653006311425",
    "1580647905458733056",
    "1580516385327050752",
    "1580433567666307072",
    "1580357002697375744",
    "1580358845402341377",
    "1580764245322792961",
    "1580356963287805952",
    "1580384464416706560",
    "1580375748228968448",
    "1580614539585781760",
    "1580514408299913216",
    "1580601818543165441",
    "1580455646193127426",
    "1580342402677878785",
    "1580341773074776065",
    "1580569567998267392",
    "1580568900630351872",
    "1581866227148161029",
    "1580644053724450816",
    "1580488252842455045",
    "1580657450008842241",
    "1580343246706049026",
    "1580411038289121281",
    "1580358697905098753",
    "1580665190084796417",
    "1580628091835420672",
    "1580690563879559168",
    "1580886628880764929",
    "1580519613523058688",
    "1580624290939998208",
    "1580345900114726912",
    "1580347272495128576",
    "1580416819852062721",
    "1580342052784480256",
    "1580399739040071680",
    "1580609815835201536",
    "1580597747530764299",
    "1580363902344253440",
    "1580670334474551296",
    "1580463526023307264",
    "1580464558459584512",
    "1580367689276497921",
    "1580745810764890112",
    "1580710346662514688",
    "1580591665731207169",
    "1580337215309512705",
    "1580370357428846592",
    "1580681056541626370",
    "1580970939646640129",
    "1580419151536877568",
    "1580394717187305472",
    "1580404140655906816",
    "1581868158293794817",
    "1581868910705401856",
    "1580389225622515712",
    "1580518657213415428",
    "1580360138808250369",
    "1580426586691899392",
    "1580370291867279360",
    "1580415821041397761",
    "1580432235148476416",
    "1580338829164101632",
    "1580588504572383232",
    "1580336897465122817",
    "1580441718788599808",
    "1580544625760051200",
    "1580353993498165248",
    "1580577468351926274",
    "1580396980928663552",
    "1580539824909750275",
    "1580465534113439744",
    "1580649674473877504",
    "1580496083935191040",
    "1580619841027051520",
    "1580540827377364995",
    "1580562742540324865",
    "1580399986684329986",
    "1580375133268869120",
    "1580654281962946560",
    "1580597013099737089",
    "1580545556325072896",
    "1580356893888823296",
    "1580387672136130560",
    "1580416864404271105"
  )

chrystia_tweets_oct_12 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_oct_12)

as_tibble(chrystia_tweets_oct_12[["tweets"]])

as_tibble(chrystia_tweets_oct_12[["users"]])

write_csv(x = chrystia_tweets_oct_12[["users"]],
          file = "raw_chrystia_tweets_oct_12.csv")

write_csv(x = chrystia_tweets_oct_12[["tweets"]],
          file = "raw_chrystia_tweets_oct_12_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/cafreeland/status/1579924784234008576",
    "1579945735826845697",
    "1579951723384537094",
    "1579977312371904512",
    "1579979637383663617",
    "1579991314950926337",
    "1580004509618941953",
    "1580004529726443522",
    "1580036054295248896",
    "1580039647329021953",
    "1580172834298810369",
    "1580263094626189317",
    "1581271378988531712",
    "1580030320367534080",
    "1580054227925499904",
    "1580090789056507904",
    "1580267626907201536",
    "1579940510550360064",
    "1580037332069384193",
    "1580016575582990336",
    "1580146134647189505",
    "1579954512206647296",
    "1579993298416336896",
    "1579946481473441792",
    "1579978906199031816",
    "1580437568025694208",
    "1580001014392164353",
    "1579934860218884096",
    "1579975449207508994",
    "1579955046749736960",
    "1579968606569660418",
    "1579935988549582848",
    "1579944269611421696",
    "1579925936015020033",
    "1579925952154701824",
    "1579926933454086145",
    "1579926952408125440",
    "1579927313734828033",
    "1579928722291175429",
    "1579929242615566336",
    "1579932623501430784",
    "1579925545428844546",
    "1579935536214847490",
    "1579936265474314241",
    "1579955777955299329",
    "1579971535028514817",
    "1579944743811043329",
    "1579938682400026624",
    "1579946160852467712",
    "1579957067322097664",
    "1579969888751587328",
    "1579994571844784128",
    "1580006333335539712",
    "1580007527835930625",
    "1580026562933448704",
    "1580030201072799746",
    "1580102973400657920",
    "1579977849821618178",
    "1580165913646870529",
    "1589421965185789953",
    "1589422710261964800",
    "1579949417976991744",
    "1579985172665360385",
    "1579987589737897985",
    "1579991225117339653",
    "1580196893287542785",
    "1580278443492913152",
    "1579927536477548545",
    "1579933057578303488",
    "1579939144629112832",
    "1579946988992593920",
    "1580043618756067328",
    # level 1 :) #
    "1588992127031443456",
    "1579974425310810112",
    "1580211444846845954",
    "1579926013878087680",
    "1579926516695437314",
    "1580018335634558976",
    "1580234248627589120",
    "1580239468769865728",
    "1579935054058631168",
    "1579942674668584960",
    "1589422226591604736",
    "1579995765661126658",
    "1580312913746997248",
    "1579925034247692288",
    "1579931529148764160",
    "1579949611019796480",
    "1580043166480498688",
    "1580161398407016449",
    "1580169996415614978",
    "1579925850090504192",
    "1579926091095224321",
    "1579928869930381312",
    "1579930382073749504",
    "1579930966411599872",
    "1579932016904396800",
    "1579933513830531072",
    "1579934375181193216",
    "1579939255094480897",
    "1579941483003928576",
    "1579950332473638914",
    "1579978619409268737",
    "1579979052471189504",
    "1579990606939496448",
    "1579994209549189120",
    "1579998591828967424",
    "1580000938349457408",
    "1580006347952619520",
    "1580011071645032448",
    "1580020370505359361",
    "1580168468472954880",
    "1580208637926322176",
    "1580216555262332933",
    "1580216665127911424",
    "1580230439884976128",
    "1580240007058853888",
    "1580243555301601281",
    "1580243930909921280",
    "1580250341110448128",
    "1580267654157590528",
    "1580273427247943682",
    "1580278416343203841",
    "1580281051494031361",
    "1580284970651615233",
    "1580285682437992449",
    "1580336699409723392",
    "1580388951356891138",
    "1580397586187358208",
    "1580661257438306306",
    "1589347838898868224"
  )

chrystia_tweets_oct_11 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_oct_11)

as_tibble(chrystia_tweets_oct_11[["tweets"]])

as_tibble(chrystia_tweets_oct_11[["users"]])

write_csv(x = chrystia_tweets_oct_11[["users"]],
          file = "raw_chrystia_tweets_oct_11.csv")

write_csv(x = chrystia_tweets_oct_11[["tweets"]],
          file = "raw_chrystia_tweets_oct_11_2.csv")

# Collect tweets, October 26 #
tweet_ids <-
  c("https://twitter.com/cafreeland/status/1717693900733751675")

chrystia_tweets_oct_26 <- tcn_threads(tweet_ids, token)

print(chrystia_tweets_oct_26)

as_tibble(chrystia_tweets_oct_26[["tweets"]])

as_tibble(chrystia_tweets_oct_26[["users"]])

write_csv(x = chrystia_tweets_oct_26[["users"]],
          file = "raw_chrystia_tweets_oct_26.csv")

write_csv(x = chrystia_tweets_oct_26[["tweets"]],
          file = "raw_chrystia_tweets_oct_26_2.csv")
