==> Scripts/01-download_Elizabeth_data.R <==
#### Preamble ####
# Purpose: Download @elizabethmay Tweets via voson.tcn
# Author: Inessa De Angelis
# Date: 4 October 2023
# Contact: inessa.deangelis@mail.utoronto.ca 
# License: MIT

#### Workspace setup ####
library(voson.tcn)
library(tidyverse)

#### Authentication ####
token <- tcn_token(bearer = "ADD SECRET CREDENTIALS")

#### Elizabeth May - key word "climate", tweets from October 2022 - October 2023 ####
# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1700787904119992692",
    "1700834322453073949",
    "1700831757996531996",
    "1700845024718594507",
    "1700825646359097434",
    "1700793175068275062",
    "1700872590565503379",
    "1700876914993561705",
    "1700899727250805232",
    "1700860101613977798",
    "1700980753155412104",
    "1700957486826615119"
  )

liz_tweets_sep10 <- tcn_threads(tweet_ids, token)

print(liz_tweets_sep10)

as_tibble(liz_tweets_sep10[["tweets"]])

as_tibble(liz_tweets_sep10[["users"]])

write_csv(x = liz_tweets_sep10[["users"]],
          file = "raw_liz_tweets_sep10.csv")

write_csv(x = liz_tweets_sep10[["tweets"]],
          file = "raw_liz_tweets_sep10_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1684608277035057152",
    "1684817729675780096",
    "1684962561442160648",
    "1684746692371910656",
    "1684626716616978432",
    "1684674776466665474",
    "1684714960897597440",
    "1684614033159094286",
    "1684667295493214209",
    "1684704429767790593",
    "1684708291765534720",
    "1684907521734250496",
    "1684766184431058944",
    "1684735271181422592",
    "1684931737636507648",
    "1685081923712946176",
    "1684668519680479232",
    "1684676561880956928",
    "1684718778993819648",
    "1684685918408749065",
    "1684612452791803904",
    "1684671239434952704",
    "1684807319744626689",
    "1685189969072287744",
    "1684737184476356609",
    "1684658878498140160",
    "1684688331316342787",
    "1684965923244130304",
    "1684965923244130304",
    "1684669353902379008",
    "1684608929438076928",
    "1685367924792463360",
    "1684888150928506880",
    "1684670969711923201",
    "1684896759204794369",
    "1684911086531674112",
    "1684677020213559296",
    "1684671594742910976",
    "1684704805858508801",
    "1684688331316342787"
  )

liz_tweets_jul_27 <- tcn_threads(tweet_ids, token)

print(liz_tweets_jul_27)

as_tibble(liz_tweets_jul_27[["tweets"]])

as_tibble(liz_tweets_jul_27[["users"]])

write_csv(x = liz_tweets_jul_27[["users"]],
          file = "raw_liz_tweets_jul_27.csv")

write_csv(x = liz_tweets_jul_27[["tweets"]],
          file = "raw_liz_tweets_jul_27_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1671243766848532484",
    "1671245134363598858",
    "1671267956251316225",
    "1671317333447671808",
    "1677000396085280770",
    "1672945825935417345"
  )

liz_tweets_jun_20 <- tcn_threads(tweet_ids, token)

print(liz_tweets_jun_20)

as_tibble(liz_tweets_jun_20[["tweets"]])

as_tibble(liz_tweets_jun_20[["users"]])

write_csv(x = liz_tweets_jun_20[["users"]],
          file = "raw_liz_tweets_jun_20.csv")

write_csv(x = liz_tweets_jun_20[["tweets"]],
          file = "raw_liz_tweets_jun_20_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1673054687086485504",
    "1673316740757958660",
    "1673330969229549570",
    "1674471642104209416",
    "1673092452176089091",
    "1673091567928889344",
    "1673060851580903425",
    "1673095745484849155",
    "1673115879800967170",
    "1673073753071456257",
    "1673074558943805442",
    "1673557951062106113",
    "1673080119458504706",
    "1673118225650061312",
    "1673805546905038849",
    "1673805691704967168",
    "1673183841836736512",
    "1673366533777805312"
  )

liz_tweets_jun_25 <- tcn_threads(tweet_ids, token)

print(liz_tweets_jun_25)

as_tibble(liz_tweets_jun_25[["tweets"]])

as_tibble(liz_tweets_jun_25[["users"]])

write_csv(x = liz_tweets_jun_25[["users"]],
          file = "raw_liz_tweets_jun_25.csv")

write_csv(x = liz_tweets_jun_25[["tweets"]],
          file = "raw_liz_tweets_jun_25_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1673800888190906368",
    "1673871169215422464",
    "1673945285381459968",
    "1673936090775048193",
    "1673824686105305089",
    "1673960620906139648",
    "1674115445769621512",
    "1674276691185741830",
    "1674040954318204929",
    "1673835030299176961",
    "1673827955968212993",
    "1673859261691207680",
    "1673976702609899524",
    "1673906632017707009",
    "1673806121889579008",
    "1674041238503440384",
    "1673938845782659074",
    "1674040930939310080",
    "1674161393795211265",
    "1673888849494441984",
    "1673905155593613313",
    "1673839692989579264",
    "1673871499328098306",
    "1673804708522389504",
    "1673819035052408832",
    "1673929522897653760",
    "1674077584135471104",
    "1673917827601039361",
    "1674033346614272003",
    "1673942623466278916",
    "1673900217601789956",
    "1673992009969696768",
    "1674037890458980352",
    "1673914383326666753",
    "1674078312597975040",
    "1673928230611439617",
    "1673949666030002177",
    "1674097092376100865",
    "1673848752304697344",
    "1673850399319818240"
  )

liz_tweets_jun_27 <- tcn_threads(tweet_ids, token)

print(liz_tweets_jun_27)

as_tibble(liz_tweets_jun_27[["tweets"]])

as_tibble(liz_tweets_jun_27[["users"]])

write_csv(x = liz_tweets_jun_27[["users"]],
          file = "raw_liz_tweets_jun_27.csv")

write_csv(x = liz_tweets_jun_27[["tweets"]],
          file = "raw_liz_tweets_jun_27_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1670615325883203584",
    "1670813808892301314",
    "1670748458842730497",
    "1671376601911803904",
    "1670801929964863495",
    "1670681151168716800",
    "1670798058676682758"
  )

liz_tweets_jun_18 <- tcn_threads(tweet_ids, token)

print(liz_tweets_jun_18)

as_tibble(liz_tweets_jun_18[["tweets"]])

as_tibble(liz_tweets_jun_18[["users"]])

write_csv(x = liz_tweets_jun_18[["users"]],
          file = "raw_liz_tweets_jun_18.csv")

write_csv(x = liz_tweets_jun_18[["tweets"]],
          file = "raw_liz_tweets_jun_18_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1668458597745868805",
    "1668488367359799296",
    "1668487794409476096",
    "1668613655225593856",
    "1668469785749778432",
    "1669064119083737096",
    "1669062131029778445",
    "1668687713040728065",
    "1668480161208205312",
    "1668509199205056512",
    "1668609177353170946",
    "1668477506947522561",
    "1668608537155739648",
    "1668554759547113473",
    "1673629227319545857",
    "1675205027047280640",
    "1668602607701467136",
    "1668543924108169216",
    "1669064970200293383",
    "1668611150651572226",
    "1668599344143138817",
    "1668732852199120896",
    "1668469165878034432"
  )

liz_tweets_jun_12 <- tcn_threads(tweet_ids, token)

print(liz_tweets_jun_12)

as_tibble(liz_tweets_jun_12[["tweets"]])

as_tibble(liz_tweets_jun_12[["users"]])

write_csv(x = liz_tweets_jun_12[["users"]],
          file = "raw_liz_tweets_jun_12.csv")

write_csv(x = liz_tweets_jun_12[["tweets"]],
          file = "raw_liz_tweets_jun_12_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1640911880582582272",
    "1641032710398279681",
    "1641050891997310978",
    "1641095598458765315",
    "1641309062523506690",
    "1641218052502036489",
    "1640975089892982784",
    "1641074751866060802",
    "1640912191191928832",
    "1641151282797740185",
    "1641009083049472000",
    "1641256083472674817",
    "1641256373139783680",
    "1641257291327365121",
    "1641257478829613057",
    "1641257785043165184",
    "1641258193828429824",
    "1641537217523752961",
    "1641117086507319311",
    "1641117286391234561",
    "1640966121544355841",
    "1641222664005226496",
    "1641170341342121984",
    "1640926435144785921",
    "1641177593348231168",
    "1641115576058912768",
    "1641163589896790016",
    "1641212988014338048",
    "1641075534657630208",
    "1641091689199595522",
    "1641223244287193088",
    "1640927823102902274"
  )

liz_tweets_mar_28 <- tcn_threads(tweet_ids, token)

print(liz_tweets_mar_28)

as_tibble(liz_tweets_mar_28[["tweets"]])

as_tibble(liz_tweets_mar_28[["users"]])

write_csv(x = liz_tweets_mar_28[["users"]],
          file = "raw_liz_tweets_mar_28.csv")

write_csv(x = liz_tweets_mar_28[["tweets"]],
          file = "raw_liz_tweets_mar_28_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1637916531098271745",
    "1637916534235619328",
    "1637916538320850944",
    "1637916539658838019",
    "1638361653308973056",
    "1638217529658056729",
    "1638361995425763328",
    "1638643925727715328",
    "1638340454348627969",
    "1638545843132284928",
    "1638360361295581184",
    "1638366335829954560",
    "1638028867045842945",
    "1638358792474873861",
    "1637945956317507585",
    "1638351280652582912",
    "1638346655534837760"
  )

liz_tweets_mar_20 <- tcn_threads(tweet_ids, token)

print(liz_tweets_mar_20)

as_tibble(liz_tweets_mar_20[["tweets"]])

as_tibble(liz_tweets_mar_20[["users"]])

write_csv(x = liz_tweets_mar_20[["users"]],
          file = "raw_liz_tweets_mar_20.csv")

write_csv(x = liz_tweets_mar_20[["tweets"]],
          file = "raw_liz_tweets_mar_20_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1625234645724852262",
    "1625250891786776602",
    "1625241359266226176",
    "1625594470484828181",
    "1625596703700316160",
    "1625238742851276803",
    "1625288346447781888",
    "1625332101674590208",
    "1625238952348188673",
    "1625578165904244736",
    "1625309981485461506",
    "1625238319109148672",
    "1632349235424428033",
    "1625253108078235650",
    "1626556018133008389",
    "1625236191263211520",
    "1625239170896769024"
  )

liz_tweets_feb_13 <- tcn_threads(tweet_ids, token)

print(liz_tweets_feb_13)

as_tibble(liz_tweets_feb_13[["tweets"]])

as_tibble(liz_tweets_feb_13[["users"]])

write_csv(x = liz_tweets_feb_13[["users"]],
          file = "raw_liz_tweets_feb_13.csv")

write_csv(x = liz_tweets_feb_13[["tweets"]],
          file = "raw_liz_tweets_feb_13_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1613585769226207234",
    "1614639661569421312",
    "1614671788667019265",
    "1614645949665804289",
    "1614683562380255232",
    "1614649730335932418",
    "1614702696799555586",
    "1614611606448529408",
    "1614602892853317632",
    "1614758783301087233",
    "1614618387967082498",
    "1614602166282649600",
    "1614931365438554117",
    "1614638394029805570",
    "1614653841286500352",
    "1614604121054785538",
    "1614872509786251264",
    "1614630835084214272",
    "1614636533398503426",
    "1614603802388172802",
    "1614620499161522177",
    "1614655461361602560",
    "1614669200965902343",
    "1614624551769411585",
    "1614625757405511681",
    "1614716542822023168",
    "1614603096839143426",
    "1614644039726534658",
    "1614615323650715650",
    "1614667017939996672",
    "1614622810755719171"
  )

liz_tweets_jan_12 <- tcn_threads(tweet_ids, token)

print(liz_tweets_jan_12)

as_tibble(liz_tweets_jan_12[["tweets"]])

as_tibble(liz_tweets_jan_12[["users"]])

write_csv(x = liz_tweets_jan_12[["users"]],
          file = "raw_liz_tweets_jan_12.csv")

write_csv(x = liz_tweets_jan_12[["tweets"]],
          file = "raw_liz_tweets_jan_12_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1604232990510522378",
    "1604281773814272000",
    "1604460922768355329",
    "1604242241224335361",
    "1604495278090129408",
    "1604357060308586496",
    "1604236648258371584",
    "1604264302340456448",
    "1604247495000850433",
    "1604400661436968960",
    "1604248258964033537",
    "1604487158852513792",
    "1604551038983626752",
    "1604329568898359299",
    "1604235568698769410"
  )

liz_tweets_dec_17 <- tcn_threads(tweet_ids, token)

print(liz_tweets_dec_17)

as_tibble(liz_tweets_dec_17[["tweets"]])

as_tibble(liz_tweets_dec_17[["users"]])

write_csv(x = liz_tweets_dec_17[["users"]],
          file = "raw_liz_tweets_dec_17.csv")

write_csv(x = liz_tweets_dec_17[["tweets"]],
          file = "raw_liz_tweets_dec_17_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1593690106501758980",
    "1594153815036645379"
  )

liz_tweets_nov_18 <- tcn_threads(tweet_ids, token)

print(liz_tweets_nov_18)

as_tibble(liz_tweets_nov_18[["tweets"]])

as_tibble(liz_tweets_nov_18[["users"]])

write_csv(x = liz_tweets_nov_18[["users"]],
          file = "raw_liz_tweets_nov_18.csv")

write_csv(x = liz_tweets_nov_18[["tweets"]],
          file = "raw_liz_tweets_nov_18_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1588566257166479361",
    "1588567292329086978",
    "1588828423551844352",
    "1588612924506927109",
    "1588853657810145280",
    "1588569178151481345"
  )

liz_tweets_nov_4 <- tcn_threads(tweet_ids, token)

print(liz_tweets_nov_4)

as_tibble(liz_tweets_nov_4[["tweets"]])

as_tibble(liz_tweets_nov_4[["users"]])

write_csv(x = liz_tweets_nov_4[["users"]],
          file = "raw_liz_tweets_nov_4.csv")

write_csv(x = liz_tweets_nov_4[["tweets"]],
          file = "raw_liz_tweets_nov_4_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1584257086761750528",
    "1584611042713571328",
    "1584520125155520512",
    "1584630474026397696",
    "1584517092099108865",
    "1584393443899772928",
    "1584619340149641216",
    "1584510699816222720",
    "1584407676330848257",
    "1584535989745160193",
    "1584617945611010048",
    "1584264767266750464"
  )

liz_tweets_oct_23 <- tcn_threads(tweet_ids, token)

print(liz_tweets_oct_23)

as_tibble(liz_tweets_oct_23[["tweets"]])

as_tibble(liz_tweets_oct_23[["users"]])

write_csv(x = liz_tweets_oct_23[["users"]],
          file = "raw_liz_tweets_oct_23.csv")

write_csv(x = liz_tweets_oct_23[["tweets"]],
          file = "raw_liz_tweets_oct_23_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1583545148981784576",
    "1583638871052812288",
    "1583575250490773515"
  )

liz_tweets_oct_21 <- tcn_threads(tweet_ids, token)

print(liz_tweets_oct_21)

as_tibble(liz_tweets_oct_21[["tweets"]])

as_tibble(liz_tweets_oct_21[["users"]])

write_csv(x = liz_tweets_oct_21[["users"]],
          file = "raw_liz_tweets_oct_21.csv")

write_csv(x = liz_tweets_oct_21[["tweets"]],
          file = "raw_liz_tweets_oct_21_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1577050971037134849",
    "1577052922684538880",
    "1577740409899741185",
    "1577153617890807808",
    "1577065063499767808",
    "1577090864655933440",
    "1577319956823293952",
    "1577325870213103617",
    "1577080888097378306",
    "1577054671902236672",
    "1577184043564961792",
    "1577052907736006656",
    "1577067118905806849",
    "1577131923469897728"
  )

liz_tweets_oct_3 <- tcn_threads(tweet_ids, token)

print(liz_tweets_oct_3)

as_tibble(liz_tweets_oct_3[["tweets"]])

as_tibble(liz_tweets_oct_3[["users"]])

write_csv(x = liz_tweets_oct_3[["users"]],
          file = "raw_liz_tweets_oct_3.csv")

write_csv(x = liz_tweets_oct_3[["tweets"]],
          file = "raw_liz_tweets_oct_3_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1576590415486103554",
    "1576638765799419904",
    "1578551884184682496",
    "1576661238582427648",
    "1577084749096849409",
    "1577251869037953024",
    "1576637100589346816",
    "1576617886659252225",
    "1576625045962518528",
    "1576762773375094788",
    "1576901646311133184",
    "1576686729016537089",
    "1576593275976183809",
    "1576681685512761345",
    "1577032966567788544",
    "1576684676240257024",
    "1576753366545625088",
    "1576666296141246466",
    "1576619498593222656",
    "1576731795839586304",
    "1576627324422098948",
    "1577131614542274560",
    "1576877321851588608",
    "1576991449673269248",
    "1577159975625187328",
    "1576724493523095552",
    "1577049409808105472",
    "1576992114394550272",
    "1576643464628154368",
    "1576721932564590592",
    "1576708452621070336",
    "1576994221357080576",
    "1576593843000340480",
    "1577026221388746752",
    "1577051799130734592",
    "1576791163469979654",
    "1576732107501580289",
    "1577036325026693120",
    "1577009584291471360",
    "1577077603202367488",
    "1576624306007597056",
    "1577113765534760960",
    "1576924631344504834",
    "1576992396813824000",
    "1576910255841193990",
    "1577030715082182656",
    "1576710954074591232",
    "1576608214560702465",
    "1577102303873892358",
    "1576762247388762112",
    "1577059138240413696",
    "1576629221777997829",
    "1576640309827895296",
    "1576727811888664577",
    "1576703075422597120",
    "1576644387697995776",
    "1576641859509637120",
    "1576602495312613382",
    "1577014244716863488",
    "1577034139907158016",
    "1577005372690219009",
    "1576912086130913285",
    "1576775605106143235",
    "1577015968718807040",
    "1576660679666282496",
    "1577013267930566656",
    "1577052108171677696",
    "1577049041523073027",
    "1577083811321348097",
    "1576667514347884544",
    "1577301783701262336",
    "1576616019032170497",
    "1577103277480558592",
    "1577089563369312256",
    "1576647561141956608",
    "1577107406919991296",
    "1577076251206225920",
    "1577302809036349440",
    "1576690457437908992",
    "1576690410864660481",
    "1577000314636357632",
    "1576764224042590209",
    "1576724645352701952",
    "1576932822341668865",
    "1576893939386585089",
    "1577057075166793729",
    "1577068155138342912",
    "1576995948068499457",
    "1576903813210210306",
    "1576651513292001280",
    "1576640216148103168",
    "1576949749831196673",
    "1576602549222014976",
    "1576720161390096384",
    "1576591998290518017",
    "1577004308507525121",
    "1576926176853229568",
    "1577018051559829504",
    "1576693049702649856",
    "1576935909357780993",
    "1576728053505744897",
    "1577101228743725056",
    "1577086917031886849"
  )

liz_tweets_oct_2 <- tcn_threads(tweet_ids, token)

print(liz_tweets_oct_2)

as_tibble(liz_tweets_oct_2[["tweets"]])

as_tibble(liz_tweets_oct_2[["users"]])

write_csv(x = liz_tweets_oct_2[["users"]],
          file = "raw_liz_tweets_oct_2.csv")

write_csv(x = liz_tweets_oct_2[["tweets"]],
          file = "raw_liz_tweets_oct_2_2.csv")

#### Elizabeth May - key word "climate", tweets from September 2022 ####
# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1575259064828829697",
    "1575538038809235467",
    "1575339382063894533",
    "1575613509718618112",
    "1575317839535382528",
    "1575470389769240579",
    "1575476283538112513",
    "1576575174669139974",
    "1575477515518447621",
    "1575639644821389312",
    "1575318039058608128",
    "1575639499903991808",
    "1576577778102984704",
    "1575530762560241664",
    "1575340684084924416",
    "1575342745555968000",
    "1575524023580360704",
    "1575512113048702977",
    "1575347299991293952",
    "1575375950753148928",
    "1575341947144781827",
    "1575341921005871104",
    "1575317684023185409",
    "1575506455658385409",
    "1575338459837140992",
    "1575678571477348352",
    "1575325497189437440",
    "1575329889087258626",
    "1575325423294103553",
    "1575348586585653248",
    "1575439010964869127",
    "1575698980717363200",
    "1575461850782760960",
    "1575676982368485377",
    "1575675615327719425",
    "1575479182355726341",
    "1575424186947932160",
    "1575267756143960064",
    "1575523427242352641",
    "1575508340473487360",
    "1575502449690873856",
    "1575340573275598848",
    "1575316405813235712",
    "1575475917799002112",
    "1575494162505670656",
    "1575367828198277120",
    "1575383833683394560",
    "1575350184497725440",
    "1575327968192499712",
    "1575360280997019649",
    "1575279295991259136",
    "1575488689936875522",
    "1575685159717720064",
    "1575488737613533186",
    "1575277576222027776",
    "1575340685322252289",
    "1575902524133249024",
    "1575277663350300672",
    "1575592600425607168",
    "1575587746907049985",
    "1575371282937413632",
    "1575363284881924096",
    "1575694391754498048",
    "1575481268586033152",
    "1575530431948402688",
    "1575753202544566272",
    "1575463216440037376",
    "1575689417326415874",
    "1575440630511173638",
    "1575643155264770048",
    "1575684885666467840",
    "1575998962397745152",
    "1575435569261412352",
    "1575701683887566849",
    "1575334421040971777",
    "1575472381388529664",
    "1575482881694113792",
    "1575337786215104512",
    "1575665014232739840",
    "1575701273416208384",
    "1575418476080074752",
    "1575271827110522880",
    "1575338639554646016",
    "1575623319101276161",
    "1575371093745049600",
    "1575702269323972610",
    "1575531427479060483",
    "1575494523693977600",
    "1575526870107537410",
    "1575539913428443136",
    "1575806914604457984",
    "1575501188522708993",
    "1575320803226288129",
    "1575573065722757123",
    "1575489914963382273",
    "1575681579120480256",
    "1575500778806333441",
    "1575334914056396801",
    "1575683200298684417",
    "1575480997155569666",
    "1575618647384784897"
  )

liz_tweets_sep_28 <- tcn_threads(tweet_ids, token)

print(liz_tweets_sep_28)

as_tibble(liz_tweets_sep_28[["tweets"]])

as_tibble(liz_tweets_sep_28[["users"]])

write_csv(x = liz_tweets_sep_28[["users"]],
          file = "raw_liz_tweets_sep_28.csv")

write_csv(x = liz_tweets_sep_28[["tweets"]],
          file = "raw_liz_tweets_sep_28_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1574730585800470528",
    "1574802081701138434",
    "1575060641089753088",
    "1574932673516425217",
    "1574803724572438533",
    "1574859551727583238",
    "1574744342354436096",
    "1574840579363672071",
    "1574865678661988352",
    "1576823863685832704",
    "1574857513547272207",
    "1574732174300917760",
    "1574888724957372416",
    "1574742263414292482",
    "1574829840691499008",
    "1574900009853919232",
    "1574975150755577856",
    "1574926031068463106",
    "1574731824886317056",
    "1574751442946097152",
    "1574821933467959296",
    "1574840913599369234",
    "1574841144151871537",
    "1574733545427578887",
    "1574756748703191041",
    "1574774329921138688",
    "1574737787328368645",
    "1574901422692642817",
    "1574806744970440704",
    "1574910649020579840",
    "1574870813693644800",
    "1574747342670020610",
    "1574847117339856913",
    "1574870534105731073"
  )

liz_tweets_sep_27 <- tcn_threads(tweet_ids, token)

print(liz_tweets_sep_27)

as_tibble(liz_tweets_sep_27[["tweets"]])

as_tibble(liz_tweets_sep_27[["users"]])

write_csv(x = liz_tweets_sep_27[["users"]],
          file = "raw_liz_tweets_sep_27.csv")

write_csv(x = liz_tweets_sep_27[["tweets"]],
          file = "raw_liz_tweets_sep_27_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1573010411879907335",
    "1574802081701138434",
    "1573169150897405954",
    "1573326301502513152",
    "1573413110987526144",
    "1573183758882471944",
    "1573141467572629504",
    "1573131190151925760",
    "1573131969004769280",
    "1573138941313970176",
    "1573143867125338112",
    "1573082605213601792",
    "1573141184389808128",
    "1573221963488956416",
    "1573133367343972352",
    "1573162572672823298",
    "1573168998883393536",
    "1573180232026427392",
    "1573139802169708544",
    "1573292953887715331",
    "1573212505295519744",
    "1573149638491783169",
    "1573138534449512450",
    "1573138534449512450",
    "1573132659877896194",
    "1573625896942141440",
    "1573083289711448064",
    "1573198266325401600",
    "1573137852636200960",
    "1573017006521339904",
    "1573033338285174784",
    "1573271884707557376",
    "1573730278325030913",
    "1573315506056359936",
    "1573313210006261760",
    "1573255688406581249",
    "1573175022944133120",
    "1573482816490307584",
    "1573430756869033984",
    "1573158075179356161",
    "1573386752685133829",
    "1573146141239263232",
    "1573273738300297217",
    "1573416398512353281",
    "1573310405409218560",
    "1573209667823882240",
    "1575127380670132225",
    "1573246256591548416",
    "1573137908294459392",
    "1573487317826019328",
    "1573181333261914112",
    "1573140089240457216",
    "1573187931891965952",
    "1573157055112019968",
    "1573273587649232896",
    "1573182200761253890",
    "1573387525305282560",
    "1573373937039572992",
    "1573348317966528514",
    "1573120188660301827",
    "1573134573298659328",
    "1573321151342051328",
    "1573167190857113601",
    "1573411547287883776",
    "1573104053470777344",
    "1573208609567408129",
    "1573295246112587776",
    "1573132192670429184",
    "1573269855645597698",
    "1573630340400582656",
    "1573102717597523969",
    "1573080894520586245",
    "1573292744147341313",
    "1573408930591805440",
    "1573386128769839107",
    "1573295428057300992",
    "1573199302977294336",
    "1573138817527459841",
    "1573382591495163904",
    "1573138337782956032",
    "1573302500651827201",
    "1573170074399440897",
    "1573169340459130881",
    "1573013098373517312",
    "1573666057348612097",
    "1573138558172413952",
    "1573140080650498049",
    "1573081651915145217",
    "1573083413082705920",
    "1573305706111201280",
    "1573330546780147713",
    "1573435804042792960",
    "1573397955603992576",
    "1573311896505171968",
    "1573143497485721601",
    "1573390946594701312",
    "1573167132560752640",
    "1573250132732149760",
    "1573312586506182656",
    "1573178012216225792",
    "1573140012027322370",
    "1573403253093855232",
    "1573444339170430976",
    "1573421675903070208",
    "1573029407169646593",
    "1573265078538969088",
    "1573397828092985344",
    "1573436213180366859",
    "1573298325696417793",
    "1573022419065118723",
    "1573146234571091968",
    "1573630357911601152",
    "1573388038704594946",
    "1573408479469240321",
    "1573284897166594048",
    "1573163838136274944",
    "1573286056988147712",
    "1573168418353827843",
    "1573182683844575232",
    "1573063525207257088",
    "1573432618015952896",
    "1573408587048976384",
    "1573139316909506565",
    "1573103658786758657",
    "1573297815857811461",
    "1573137112555454464",
    "1573188912675430400",
    "1573276434151649286",
    "1573137768292945921",
    "1573014067958386688",
    "1573245130743902208",
    "1573135609249792000",
    "1573302372264448007",
    "1573478836997849089",
    "1573308818007937025",
    "1573011316302053377",
    "1573294765076287488",
    "1573127524732882949",
    "1573156230969069570",
    "1573300967361183746",
    "1573140807145902080",
    "1573026611649474560",
    "1573295294250422274",
    "1573383278748635136",
    "1573295449771220998"
  )

liz_tweets_sep_22 <- tcn_threads(tweet_ids, token)

print(liz_tweets_sep_22)

as_tibble(liz_tweets_sep_22[["tweets"]])

as_tibble(liz_tweets_sep_22[["users"]])

write_csv(x = liz_tweets_sep_22[["users"]],
          file = "raw_liz_tweets_sep_22.csv")

write_csv(x = liz_tweets_sep_22[["tweets"]],
          file = "raw_liz_tweets_sep_22_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1572681638345072640",
    "1572870281538015234",
    "1572809134524841984",
    "1572719652400152577",
    "1572915529815396352",
    "1572696571632033792",
    "1572692712629104640",
    "1572686534666776577",
    "1572924231264129024",
    "1572696719493853184",
    "1573008440183635969",
    "1572687429663858688",
    "1572940982420901889",
    "1572709386627452928",
    "1572739650284027904",
    "1572722481248735236",
    "1573094718858461184",
    "1573305735278362624",
    "1572733627473203202",
    "1572688689725308928",
    "1573034971518103554",
    "1572715400634441728",
    "1572940094981677056",
    "1572831880315310080"
  )

liz_tweets_sep_21 <- tcn_threads(tweet_ids, token)

print(liz_tweets_sep_21)

as_tibble(liz_tweets_sep_21[["tweets"]])

as_tibble(liz_tweets_sep_21[["users"]])

write_csv(x = liz_tweets_sep_21[["users"]],
          file = "raw_liz_tweets_sep_21.csv")

write_csv(x = liz_tweets_sep_21[["tweets"]],
          file = "raw_liz_tweets_sep_21_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1570458932950487043",
    "1570461369744654336",
    "1570461468897968131"
  )

liz_tweets_sep_15 <- tcn_threads(tweet_ids, token)

print(liz_tweets_sep_15)

as_tibble(liz_tweets_sep_15[["tweets"]])

as_tibble(liz_tweets_sep_15[["users"]])

write_csv(x = liz_tweets_sep_15[["users"]],
          file = "raw_liz_tweets_sep_15.csv")

write_csv(x = liz_tweets_sep_15[["tweets"]],
          file = "raw_liz_tweets_sep_15_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1568711030426976256",
    "1568713852719697921"
  )

liz_tweets_sep_10 <- tcn_threads(tweet_ids, token)

print(liz_tweets_sep_10)

as_tibble(liz_tweets_sep_10[["tweets"]])

as_tibble(liz_tweets_sep_10[["users"]])

write_csv(x = liz_tweets_sep_10[["users"]],
          file = "raw_liz_tweets_sep_10.csv")

write_csv(x = liz_tweets_sep_10[["tweets"]],
          file = "raw_liz_tweets_sep_10_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1568361954313633794",
    "1568362246685982720",
    "1568417388026294273",
    "1568665219496288257",
    "1569363311086682117",
    "1568494329802661888",
    "1568567752415219712",
    "1568371266859716608",
    "1568651281472065536",
    "1568687814560350209"
  )

liz_tweets_sep_9 <- tcn_threads(tweet_ids, token)

print(liz_tweets_sep_9)

as_tibble(liz_tweets_sep_9[["tweets"]])

as_tibble(liz_tweets_sep_9[["users"]])

write_csv(x = liz_tweets_sep_9[["users"]],
          file = "raw_liz_tweets_sep_9.csv")

write_csv(x = liz_tweets_sep_9[["tweets"]],
          file = "raw_liz_tweets_sep_9_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1714032060208132347",
    "1714117235596464399",
    "1714082414186864721",
    "1714121861728977071",
    "1714385216905925029",
    "1714331102960386364",
    "1714120766193819782",
    "1714297585266036943"
  )

liz_tweets_oct_16 <- tcn_threads(tweet_ids, token)

print(liz_tweets_oct_16)

as_tibble(liz_tweets_oct_16[["tweets"]])

as_tibble(liz_tweets_oct_16[["users"]])

write_csv(x = liz_tweets_oct_16[["users"]],
          file = "raw_liz_tweets_oct_16.csv")

write_csv(x = liz_tweets_sep_9[["tweets"]],
          file = "raw_liz_tweets_oct_16_2.csv")

#### Elizabeth May - key word "environment", tweets from October 2022 - October 2023 ####
# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1653219495219953670",
    "1653318394207432705",
    "1653278085322530816",
    "1653221336066691072"
  )

liz_tweets_may_1 <- tcn_threads(tweet_ids, token)

print(liz_tweets_may_1)

as_tibble(liz_tweets_may_1[["tweets"]])

as_tibble(liz_tweets_may_1[["users"]])

write_csv(x = liz_tweets_may_1[["users"]],
          file = "raw_liz_tweets_may_1.csv")

write_csv(x = liz_tweets_may_1[["tweets"]],
          file = "raw_liz_tweets_may_1_2.csv")

#### Elizabeth May - key word "environmental", tweets from October 2022 - October 2023 #####
# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1714755356876763459",
    "1714755653405696068",
    "1714757693150277948",
    "1714852205587739097",
    "1714872150891672008"
  ) # Check back before Nov 3rd on this series #

liz_tweets_oct_18 <- tcn_threads(tweet_ids, token)

print(liz_tweets_oct_18)

as_tibble(liz_tweets_oct_18[["tweets"]])

as_tibble(liz_tweets_oct_18[["users"]])

write_csv(x = liz_tweets_oct_18[["users"]],
          file = "raw_liz_tweets_oct_18.csv")

write_csv(x = liz_tweets_oct_18[["tweets"]],
          file = "raw_liz_tweets_oct_18_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1712860672227504192",
    "1712868617451192421",
    "1712964694476988706",
    "1713014207073914999",
    "1712861215251427668",
    "1712985995648364805",
    "1713016711837040789",
    "1712917031987585426",
    "1712879974926242222",
    "1712972012035977712",
    "1712986313220059202",
    "1712914437735141467",
    "1713199094594666521",
    "1712905282215350571",
    "1712972502690119910",
    "1712864732724797592",
    "1712898931783618747",
    "1712863059503727082",
    "1713217636484460881",
    "1713177980350947838",
    "1712994521410064534",
    "1712949116081234012",
    "1713670310892073258",
    "1712893465699536993",
    "1713200155787489586",
    "1712928804270960654",
    "1713015777530700278",
    "1712938097854308772",
    "1712985209585664263",
    "1712905289656320177",
    "1713203504976892058",
    "1712925496906785122",
    "1713195186828841232",
    "1712966201901740108",
    "1713185909573648395",
    "1712972868504457304",
    "1712911505321058668",
    "1712993213370282193",
    "1712985132322492667",
    "1713484596531118586",
    "1713216644263051531",
    "1713184742827598013",
    "1712879811809759593",
    "1712864490667286704",
    "1713364500471021929",
    "1713195817463349308",
    "1713364826443882531"
  )

liz_tweets_oct_13 <- tcn_threads(tweet_ids, token)

print(liz_tweets_oct_13)

as_tibble(liz_tweets_oct_13[["tweets"]])

as_tibble(liz_tweets_oct_13[["users"]])

write_csv(x = liz_tweets_oct_13[["users"]],
          file = "raw_liz_tweets_oct_13.csv")

write_csv(x = liz_tweets_oct_13[["tweets"]],
          file = "raw_liz_tweets_oct_13_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1639017955747934209",
    "1639251602765938688",
    "1639298187226755072",
    "1639215798718959616",
    "1639018343192727553",
    "1639359171903512579"
  )

liz_tweets_mar_23 <- tcn_threads(tweet_ids, token)

print(liz_tweets_mar_23)

as_tibble(liz_tweets_mar_23[["tweets"]])

as_tibble(liz_tweets_mar_23[["users"]])

write_csv(x = liz_tweets_mar_23[["users"]],
          file = "raw_liz_tweets_mar_23.csv")

write_csv(x = liz_tweets_mar_23[["tweets"]],
          file = "raw_liz_tweets_mar_23_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1641159830999056398",
    "1641163798513070080",
    "1641160821962072064",
    "1641179549865066502",
    "1641248660263403523",
    "1641224847148863488",
    "1641160127083339777",
    "1641161101776756737",
    "1641462185447800832"
  )

liz_tweets_mar_29 <- tcn_threads(tweet_ids, token)

print(liz_tweets_mar_29)

as_tibble(liz_tweets_mar_29[["tweets"]])

as_tibble(liz_tweets_mar_29[["users"]])

write_csv(x = liz_tweets_mar_29[["users"]],
          file = "raw_liz_tweets_mar_29.csv")

write_csv(x = liz_tweets_mar_29[["tweets"]],
          file = "raw_liz_tweets_mar_29_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1641162621670834176",
    "1641188252609642496",
    "1641209557019160576",
    "1641223550613979137",
    "1641183543555416065",
    "1641463197801791490"
  )

liz_tweets_mar_29_2 <- tcn_threads(tweet_ids, token)

print(liz_tweets_mar_29_2)

as_tibble(liz_tweets_mar_29_2[["tweets"]])

as_tibble(liz_tweets_mar_29_2[["users"]])

write_csv(x = liz_tweets_mar_29_2[["users"]],
          file = "raw_liz_tweets_mar_29_3.csv")

write_csv(x = liz_tweets_mar_29_2[["tweets"]],
          file = "raw_liz_tweets_mar_29_4.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/ElizabethMay/status/1588654495332204544",
    "1588978492246327296",
    "1589031320587767808",
    "1589043460497354752",
    "1588984103038484480",
    "1588979013598711809",
    "1589316270977351680",
    "1589058873499213825",
    "1588964212818444288",
    "1588683606284988418",
    "1588685954768056322",
    "1588971562442846208",
    "1623653361563193344",
    "1589002299116257280",
    "1589052463600263171",
    "1588994110165114881",
    "1588981633092579328",
    "1589009020425474048",
    "1588993322819129344",
    "1588711248778366977",
    "1588981344243429376",
    "1588955798805565440",
    "1589007876185796608",
    "1589097167838646272",
    "1588981838164656130",
    "1588965772835303424",
    "1588957169583820800",
    "1589051397789863937",
    "1588977155924713473",
    "1588968991024107520",
    "1588895992921063424",
    "1588958346396454912",
    "1589007528415068161",
    "1588958164770512898",
    "1589018013428486145",
    "1589001812560207873",
    "1589271977386446850",
    "1588988647285870592",
    "1588978863111274496",
    "1588949202465619968",
    "1589001757375758336",
    "1589278025153417216",
    "1588965887306268672",
    "1589108079131779072",
    "1588988255449808896",
    "1588988919961776128",
    "1588976267680178176",
    "1589008049041444864",
    "1588892867401224196",
    "1588960501668929536",
    "1588964700967956480",
    "1588966852528459777",
    "1588966745423122438",
    "1588970896194433025",
    "1589118171172462592",
    "1589004137643257856",
    "1589150498556608513",
    "1589218065027588096",
    "1588960827172073472",
    "1588971553320243200",
    "1589008354458087424",
    "1589019820590534656",
    "1588981772574150656",
    "1588976717590560768",
    "1588998084918321154",
    "1588685774652084224",
    "1588974564180393984",
    "1588961754431692800",
    "1588951488147386368",
    "1588965136953638912",
    "1588958642644332545",
    "1588915936614318080",
    "1588995207974256640",
    "1589082867745054721",
    "1588986618282901504",
    "1588993924298723328",
    "1588992895327272960",
    "1588950341932179456",
    "1588997221919625216",
    "1588987050510155776",
    "1589300200799367168",
    "1589444888386756610",
    "1589033273371787264",
    "1588967524418596867",
    "1588969400417538053",
    "1589281543515942913",
    "1589012350920974337",
    "1588996812371025921",
    "1589000752005263360",
    "1588976565203144705",
    "1588951056440258560",
    "1589017525345726464",
    "1589059780156743680",
    "1588974523520798720",
    "1588997937644068864",
    "1623467151993745409",
    "1589003060755697665",
    "1588987665579642880",
    "1588992926855876608",
    "1588969558416977920",
    "1588955941915230210",
    "1588959233424625664",
    "1589001700849119232",
    "1589337078369615872",
    "1623524869555253248",
    "1623486590377918466",
    "1588971145650647040",
    "1588989648235532288",
    "1589260509324664832",
    "1589284345784991744",
    "1588973032781262848",
    "1589006726107631616",
    "1588949880130924544",
    "1588993118208421888",
    "1589024671584116736",
    "1589105270508032000",
    "1588986779247738880",
    "1588959324638179330",
    "1588976717242433536",
    "1589058364537176066",
    "1589007152307658754",
    "1588949908572491776",
    "1589038153226399748",
    "1588981968095825920",
    "1588999936259293184",
    "1589215365967400960",
    "1589107793264795649",
    "1588981764974063617",
    "1589262070041960453",
    "1589125255637127168",
    "1589074289969754113",
    "1588988584438403072",
    # Level 7, yikes #
    "1588955061283352576",
    "1588952113396453377",
    "1589028055477256194",
    "1588972866590367745",
    "1588964856098848770",
    "1588968920987635712",
    "1588958802350837760",
    "1588988321573015552",
    "1588988365789368321",
    "1589070934539112450",
    "1588977815944192000",
    "1589004297769193473",
    "1588974254867247105",
    "1589021820534984704",
    "1588989705538146305",
    "1588721186284142593",
    "1588978786129043456",
    "1588967705024991232",
    "1589146862371512320",
    "1589002121479090176",
    "1588973503696732160",
    "1589016921479213056",
    "1589015107023310848",
    "1588990011340644352",
    "1588999562060259329",
    "1589056427431428096",
    "1588965109090910211"
  )

liz_tweets_no_4 <- tcn_threads(tweet_ids, token)

print(liz_tweets_no_4)

as_tibble(liz_tweets_no_4[["tweets"]])

as_tibble(liz_tweets_no_4[["users"]])

write_csv(x = liz_tweets_no_4[["users"]],
          file = "raw_liz_tweets_no_4.csv")

write_csv(x = liz_tweets_no_4[["tweets"]],
          file = "raw_liz_tweets_no_4_2.csv")

# *NEW* Collect tweets #
tweet_ids <-
  c("https://twitter.com/ElizabethMay/status/1717737226615070991")

liz_tweets_oct_26 <- tcn_threads(tweet_ids, token)

print(liz_tweets_oct_26)

as_tibble(liz_tweets_oct_26[["tweets"]])

as_tibble(liz_tweets_oct_26[["users"]])

write_csv(x = liz_tweets_oct_26[["users"]],
          file = "raw_liz_tweets_oct_26.csv")

write_csv(x = liz_tweets_oct_26[["tweets"]],
          file = "raw_liz_tweets_oct_26_2.csv")
