==> Scripts/01-download_Laurel_data.R <==
#### Preamble ####
# Purpose: Download @laurel_bc Tweets via voson.tcn
# Author: Inessa De Angelis
# Date: 4 October 2023
# Contact: inessa.deangelis@mail.utoronto.ca 
# License: MIT

#### Workspace setup ####
library(voson.tcn)
library(tidyverse)

#### Authentication ####
token <- tcn_token(bearer = "ADD SECRET CREDENTIALS")

#### Laurel Collins - key word "climate", tweets from October 2022 - October 2023 #####
# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1709351908400562457",
    "1709596434259984455",
    "1710320459806331123"
  )

laurel_tweets_oct_3 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_oct_3)

as_tibble(laurel_tweets_oct_3[["tweets"]])

as_tibble(laurel_tweets_oct_3[["users"]])

write_csv(x = laurel_tweets_oct_3[["users"]],
          file = "raw_laurel_tweets_oct_3.csv")

write_csv(x = laurel_tweets_oct_3[["tweets"]],
          file = "raw_laurel_tweets_oct_3_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1692249500797264227",
    "1692249755882319980",
    "1692575263790448735",
    "1692412069466005800",
    "1692569627749224907",
    "1692384447012708395",
    "1692365254309687385",
    "1692480797549494510",
    "1692555897036243248",
    "1692569432453775840",
    "1692572482786734501",
    "1692267562552418795",
    "1692572723694780590",
    "1692504561028661636",
    "1692581199162818798",
    "1692359970564780112",
    "1692560268046545126",
    "1692382056565940401",
    "1692570379166925164",
    "1692367027581984927",
    "1692409270170706038",
    "1692610187046146396",
    "1692565656992936098",
    "1692544063390830945",
    "1692562203059142697",
    "1692695063166439861",
    "1692562611609407771",
    "1692566623193084261",
    "1692563780847141003",
    "1692563805144772941",
    "1692934016054259883",
    "1692562613962416605",
    "1693788586347491625",
    "1692348327898009766",
    "1692563240465678623",
    "1692568636748357964",
    "1692382919439102026",
    "1692383023311061486",
    "1692646397873623454",
    "1692431998114316417",
    "1692530259881783371",
    "1692578969294655598",
    "1692523232379756927",
    "1692585126599725523",
    "1692595960977752104",
    "1692474682941775913",
    "1692574409087959285",
    "1692553595965849967",
    "1692379844846276824",
    "1692565375903248417",
    "1692569851812889058",
    "1692558510750351787",
    "1692350483237601532",
    "1692564426195706020",
    "1692469882019791009",
    "1692391667079270872",
    "1692507848024932441",
    "1692559172427014457",
    "1692282735224864796",
    "1692557873715245485",
    "1692429452406603986",
    "1692378749877317648",
    "1692571120027812313",
    "1692567883077873837",
    "1692564136499310712",
    "1692616362034880800",
    "1692562184352723350",
    "1692396273716871177",
    "1692612644098707604",
    "1692564075526635700",
    "1692422748394434976",
    "1692561915430723710",
    "1692601592372191340",
    "1692552668881854740",
    "1692559705267392592",
    "1692580981742403592",
    "1692579903387127853",
    "1692564281722884499",
    "1692371844311965958",
    "1692370540441305141",
    "1692567409457307854",
    "1692562070175416415",
    "1692369419740287284",
    "1692585400035058068",
    "1692405153910149175",
    "1692255575172866176",
    "1692890408722850102",
    "1692602277520195841",
    "1692560726823494079",
    "1692579712852484472",
    "1692562290174783659",
    "1692579962505539906",
    "1692567787355377944",
    "1692585161635066349",
    "1692560370995532256",
    "1692347376315351514",
    "1692560897355690428",
    "1692265742270619918",
    "1692414003090518406",
    "1692602952463135225",
    "1692956491337081236",
    "1692616647322829016",
    "1692674784188940644",
    "1692677287810834748",
    "1693313117034320131",
    "1692386015837827392",
    "1692561651042533387",
    "1692562726592143741",
    "1692568389125042548",
    "1692566727098245553",
    "1692561095947399651",
    "1692569229600436589",
    "1692564169277517898",
    "1692563261193961961",
    "1692399359072309518",
    "1692571593384530182",
    "1692572386426761661",
    "1692571488577352001",
    "1692565288544313480",
    "1692570667609211146",
    "1692598440222490690",
    "1692566054705537247",
    "1692561600404660474",
    "1692647601592574343",
    "1692504799923535931",
    "1692516975610593501",
    "1692568559199912421",
    "1692296278586343863",
    "1692560785099174303",
    "1692609940634870006",
    "1692556411450872198",
    "1692562057663787518",
    "1692273401937944669",
    "1692273863726649566",
    "1692562944003826131",
    "1692586258604699887",
    "1692573365356183862",
    "1692404426815033533",
    "1692363891098218896",
    "1692366914654535816",
    "1692569462522757556",
    "1692529962891255834",
    "1692545402002714949",
    "1692702769839562904",
    "1692702857773150308",
    "1692561710731710639",
    "1692383624992264477",
    "1692563236493668855",
    "1692594944706260993",
    "1692584805420892588",
    "1692584969208537376",
    "1692585210984931461",
    "1692403022968815970",
    "1692598482735964387",
    "1692610482400813310",
    "1692406945360285822",
    "1692563034345279514",
    "1692561276164354162",
    "1692348889255538714",
    "1692584103147901410",
    "1692386286764708225",
    "1692568995931820454",
    "1692570982320423073",
    "1692582477330514128",
    "1692564678038176240",
    "1692573828709335084",
    "1692570017852961140",
    "1692562147069608120",
    "1692559862239318117",
    "1692560924932968802",
    "1692562880174960993",
    "1692646575674159174",
    "1692569505241747752",
    "1692355519074513235",
    "1692578732404359292",
    "1693126727667696000",
    "1692358025703759944",
    "1692557668286902651",
    "1692412219588550745",
    "1692560146063339939",
    "1692600378905264273",
    "1692395211031535679",
    "1692575158068822350",
    "1692531334487441476",
    "1692531683872956429",
    "1692595120376352928",
    "1692563630246810092",
    "1692744091271115175",
    "1692390853338398935",
    "1692390897416282214"
  )

laurel_tweets_aug_17 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_aug_17)

as_tibble(laurel_tweets_aug_17[["tweets"]])

as_tibble(laurel_tweets_aug_17[["users"]])

write_csv(x = laurel_tweets_aug_17[["users"]],
          file = "raw_laurel_tweets_aug_17.csv")

write_csv(x = laurel_tweets_aug_17[["tweets"]],
          file = "raw_laurel_tweets_aug_17_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1672299255107825664",
    "1672299363643830272",
    "1672630976990568451",
    "1672348659365838848",
    "1672506285944602625",
    "1672334120456957957",
    "1672313820772175872",
    "1672809121064849408",
    "1674605091247226881",
    "1672346410430377985",
    "1672322331254063108",
    "1672317705733414925",
    "1672317554545790976",
    "1672368026581180417",
    "1672335684416466944",
    "1672320685388201984",
    "1672318868218613762",
    "1672581217290469376",
    "1672344875659063296",
    "1672348651732041731",
    "1672640457744973826",
    "1672640581703540738",
    "1672333338202472448",
    "1672405576297615360",
    "1672579923964317696",
    "1672318975785697280",
    "1672346676592295936",
    "1672355780039680001",
    "1672322004530364418",
    "1672331748514881539",
    "1672325651586744328",
    "1672402909471084545",
    "1672310774130434048",
    "1672377046012403712",
    "1672413442723176448",
    "1672323653181095938",
    "1672612845031419905",
    "1672773363645952000",
    "1672322857874096129",
    "1672358982080679936",
    "1672383953964244993",
    "1672316479201763328",
    "1672437893753901056",
    "1672567446518083585",
    "1672373205477265408",
    "1672747260986228741",
    "1672323539926646787",
    "1672421691195097089",
    "1672324585390329856",
    "1672377554471362566",
    "1672328203137093632",
    "1672346189390557185",
    "1672419583708499968",
    "1672309910741323776",
    "1672320265940791296",
    "1672591707282579457",
    "1672592409912307713",
    "1672393169420406784",
    "1672393228786577409",
    "1672393312479657985",
    "1672361335861006336",
    "1672500845215350784",
    "1672727013369720832",
    "1672355476359778305",
    "1672417939222568960",
    "1672692780253208576",
    "1672470667525079041",
    "1672326886918324226",
    "1672351494560190464",
    "1672381643179577344",
    "1672402754025791489",
    "1672400809361920003",
    "1672428669707354112",
    "1672324276542898182",
    "1672369671566409730",
    "1672371187022954496",
    "1672327576457715712",
    "1672324797630648321",
    "1672407199107231745",
    "1672319202018091008",
    "1672327359100485639",
    "1672367086449090561",
    "1672369621352198144",
    "1672420017525194753",
    "1672308589715918849",
    "1672424041091850240",
    "1672615949734924288"
  )

laurel_tweets_jun_23 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_jun_23)

as_tibble(laurel_tweets_jun_23[["tweets"]])

as_tibble(laurel_tweets_jun_23[["users"]])

write_csv(x = laurel_tweets_jun_23[["users"]],
          file = "raw_laurel_tweets_jun_23.csv")

write_csv(x = laurel_tweets_jun_23[["tweets"]],
          file = "raw_laurel_tweets_jun_23_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1668321332319375361",
    "1668615793662099462",
    "1669951073203933184",
    "1669440821005897750",
    "1668341896882819072",
    "1668589785324085248",
    "1668630774508326912",
    "1668637243991359488",
    "1668634271849025537",
    "1668608026583126016",
    "1668653721650331650",
    "1668612113336549378",
    "1668651863699509249",
    "1668596832023216133",
    "1668585527908417536",
    "1668592749254434816",
    "1668607974305062912",
    "1668653263225339905",
    "1668576513170472960",
    "1668594687681126401",
    "1668684683616587776",
    "1668657027097444352",
    "1668634417554939909",
    "1668642191818850304",
    "1668614180654706690",
    "1668580541912981506",
    "1668598023172280320",
    "1668589103259037697",
    "1668651084133400577",
    "1668602640073396225",
    "1668588616887463937",
    "1668617854697226240",
    "1668640891878395907",
    "1668585117126721536",
    "1668646751115579392",
    "1668615157012987904",
    "1668682836545765376",
    "1668672535968518145",
    "1668673039809261569",
    "1668390530022273024",
    "1668582604038340610",
    "1668637121702252546",
    "1668730734432112640",
    "1668607986661486595",
    "1668577142739746819",
    "1668508391118082048",
    "1668657695711277068",
    "1668641289594871811"
  )

laurel_tweets_jun_12 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_jun_12)

as_tibble(laurel_tweets_jun_12[["tweets"]])

as_tibble(laurel_tweets_jun_12[["users"]])

write_csv(x = laurel_tweets_jun_12[["users"]],
          file = "raw_laurel_tweets_jun_12.csv")

write_csv(x = laurel_tweets_jun_12[["tweets"]],
          file = "raw_laurel_tweets_jun_12_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1666457807040794624",
    "1666688424437518336",
    "1666572095848079360",
    "1666489002738737175",
    "1666768822253260800",
    "1666592781442363393"
  )

laurel_tweets_jun_7 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_jun_7)

as_tibble(laurel_tweets_jun_7[["tweets"]])

as_tibble(laurel_tweets_jun_7[["users"]])

write_csv(x = laurel_tweets_jun_7[["users"]],
          file = "raw_laurel_tweets_jun_7.csv")

write_csv(x = laurel_tweets_jun_7[["tweets"]],
          file = "raw_laurel_tweets_jun_7_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1649928349714911232",
    "1649991341559156736",
    "1649953849233522688",
    "1650165790895923201",
    "1650195498287128582",
    "1649944844498763781",
    "1649945950612512768",
    "1650136289260326912",
    "1649985452450119680",
    "1650166087059931136",
    "1650165620280037376",
    "1650166178298605569",
    "1650001290389499904",
    "1649942225785323520",
    "1650155511483088896",
    "1649990565554180096",
    "1650251028397187072",
    "1649995987505782784",
    "1650222851117424641",
    "1650094488415621120",
    "1650169930493599750",
    "1650129034712424449",
    "1650238187308982272",
    "1650137199026003971",
    "1649949470485020675",
    "1650013702433939456",
    "1650187029320454146"
  )

laurel_tweets_apr_22 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_apr_22)

as_tibble(laurel_tweets_apr_22[["tweets"]])

as_tibble(laurel_tweets_apr_22[["users"]])

write_csv(x = laurel_tweets_apr_22[["users"]],
          file = "raw_laurel_tweets_apr_22.csv")

write_csv(x = laurel_tweets_apr_22[["tweets"]],
          file = "raw_laurel_tweets_apr_22_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1646175281189896192",
    "1646337606190501889",
    "1646220940651663360",
    "1646206158510882817",
    "1646345586633760769",
    "1646206354963546180",
    "1646495629588434949",
    "1646236968949252097",
    "1646320136138420224",
    "1646353775039500288",
    "1646343687818067968",
    "1646303036904738817",
    "1646323493825310720",
    "1646491768391704577",
    "1646300823197196289",
    "1646213931726409728",
    "1646212807820734465",
    "1646260802847129600",
    "1646605165431238656",
    "1646228005008838656"
  )

laurel_tweets_apr_12 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_apr_12)

as_tibble(laurel_tweets_apr_12[["tweets"]])

as_tibble(laurel_tweets_apr_12[["users"]])

write_csv(x = laurel_tweets_apr_12[["users"]],
          file = "raw_laurel_tweets_apr_12.csv")

write_csv(x = laurel_tweets_apr_12[["tweets"]],
          file = "raw_laurel_tweets_apr_12_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1628209115741966337",
    "1628471579209244674",
    "1628474085700407296",
    "1628297436690272258",
    "1628251836980273153",
    "1628251836980273153",
    "1628386451430969345"
  )

laurel_tweets_feb_21 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_feb_21)

as_tibble(laurel_tweets_feb_21[["tweets"]])

as_tibble(laurel_tweets_feb_21[["users"]])

write_csv(x = laurel_tweets_feb_21[["users"]],
          file = "raw_laurel_tweets_feb_21.csv")

write_csv(x = laurel_tweets_feb_21[["tweets"]],
          file = "raw_laurel_tweets_feb_21_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1625563670100213760",
    "1625567468063432705",
    "1625672116409630726",
    "1625878622245621760"
  )

laurel_tweets_feb_14 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_feb_14)

as_tibble(laurel_tweets_feb_14[["tweets"]])

as_tibble(laurel_tweets_feb_14[["users"]])

write_csv(x = laurel_tweets_feb_14[["users"]],
          file = "raw_laurel_tweets_feb_14.csv")

write_csv(x = laurel_tweets_feb_14[["tweets"]],
          file = "raw_laurel_tweets_feb_14_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1624127211006738433",
    "1624356238959030273",
    "1624437334983983105",
    "1624463489090592769",
    "1624406961490997253",
    "1624507193499258881",
    "1624402297940807682",
    "1624290646646222848",
    "1624412567039799298",
    "1624418270811148288",
    "1624154423705600000",
    "1624374252035006466",
    "1624451580526493698",
    "1624368718540156930",
    "1624401878141517825",
    "1624504452358225920",
    "1624381829040988160",
    "1624432033966563330",
    "1624483224977133576",
    "1624247552379072512",
    "1624302050715791362",
    "1624179934074662912",
    "1624622373637857280",
    "1624249482199330816",
    "1624409657371422724",
    "1624417517425905668",
    "1624467227301888001",
    "1624220362526580736",
    "1624216699607347202",
    "1624353652151570432",
    "1625052474447912960",
    "1624507202672394241",
    "1624958935659606017",
    "1624269193054150659",
    "1624420813314359296",
    "1624514080781344768",
    "1624690563852148736",
    "1624358669893750784",
    "1624300637302280192",
    "1624457493064867841",
    "1624281839191064576",
    "1624283371223519232",
    "1624541932989865986",
    "1624409300985683970",
    "1624372219739619328",
    "1624223651150303232",
    "1624281664074424322",
    "1624306147766472704",
    "1624294378016747524",
    "1624298743502716930",
    "1624501072952803328",
    "1624418138157912064",
    "1624376947353452545",
    "1624435143112007682",
    "1624292918617083904",
    "1624427109732225024",
    "1624399609240989701",
    "1624428179149692928",
    "1624401158658719744",
    "1624473800116178944",
    "1624411031857504258",
    "1624548060159422464",
    "1624529207144648708",
    "1624525196034990081",
    "1624452378371006468",
    "1624513276272054272",
    "1624282115331469312",
    "1624319330954674178",
    "1624431756073132033",
    "1624309077999517696",
    "1624390402487861253",
    "1624411988322394113",
    "1624426905163165699",
    "1624419995982725120",
    "1624415632404905984",
    "1624384089116622854",
    "1624435277296021504",
    "1624407800586842112",
    "1624385363786493952",
    "1624294335373279232",
    "1624469515726733324",
    "1624469515726733324",
    "1624423609375813632",
    "1624406575145254914",
    "1624445272188387328",
    "1624348986214694912",
    "1624243710505656322",
    "1624448694119550978",
    "1624454847381770240",
    "1624451988087001088",
    "1624408346026450944",
    "1624427892129030144",
    "1624445852793331712",
    "1624743430222929921",
    "1624542191073869824",
    "1624514350953406465",
    "1624411206416039938",
    "1624720847645683712",
    "1624518807602737152",
    "1624407991393914884",
    "1624431911937662976",
    "1624436008598261762",
    "1624416283575058432",
    "1624488464547815425",
    "1624448780895657984",
    "1624448826387107840",
    "1624474787929260032",
    "1624439632493510657",
    "1624479957752635392",
    "1624475185817460738",
    "1624417473767538689",
    "1624273977656102915",
    "1624462384751419392",
    "1624395981188997123",
    "1624456904234655745",
    "1624449481956634628",
    "1624481153825615872",
    "1624419914693083137",
    "1624410112918953984",
    "1624412920166641667",
    "1624445458251943936",
    "1624409644369059843",
    "1624366951807045632",
    "1624424229927288832",
    "1624437665184747520",
    "1624464534348959747",
    "1624771371971579904",
    "1624357456129282048",
    "1624409443344547842",
    "1624486650066898944",
    "1624421157012373505",
    "1624530150032543744",
    "1624301533801377793",
    "1624300000602882050",
    "1624566739815002112",
    "1624429652654653444",
    "1624593994633322497",
    "1624407732085223426",
    "1624315459406548992",
    "1624399525476724736",
    "1624382753851441153",
    "1624421197139308544",
    "1624441668832223232",
    "1624400169319165952",
    "1624467700041895937",
    "1624465595235467264",
    "1624811044257566720",
    "1624492214855380997",
    "1624456682222022656",
    "1624369684861669377",
    "1624570008822226944",
    "1624468058676006912",
    "1624649593450291200",
    "1624590131612250112",
    "1624481355764535296",
    "1624465126719107074",
    "1624466727143886848",
    "1624402980698050562",
    "1624496661971226624",
    "1624520182726275072",
    "1624383595480510466",
    "1624537545487925249",
    "1624521111328198660",
    "1624423667902910470",
    "1624475193287770119",
    "1624433662770188288",
    "1624492876184039425",
    "1624417033143087104",
    "1624437395247759362",
    "1624346965860405249",
    "1624445898657767426",
    "1624492210237636609",
    "1624466096958078977",
    "1624453568681095168",
    "1624448118375817216",
    "1624395647905308674",
    "1624404449228734466",
    "1624500158338859008",
    "1624464589453664256",
    "1624487115139518467",
    "1624390128822112256",
    "1624813254814859265",
    "1624715839961464832",
    "1624611609875255298",
    "1624347260472508416",
    "1624449865424986123",
    "1624510811975680006",
    "1624292903882264577",
    "1624525125017292800",
    "1624284623030353920",
    "1624376059448774656"
  )

laurel_tweets_feb_10 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_feb_10)

as_tibble(laurel_tweets_feb_10[["tweets"]])

as_tibble(laurel_tweets_feb_10[["users"]])

write_csv(x = laurel_tweets_feb_10[["users"]],
          file = "raw_laurel_tweets_feb_10.csv")

write_csv(x = laurel_tweets_feb_10[["tweets"]],
          file = "raw_laurel_tweets_feb_10_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1620520104491634688",
    "1620543343913766912",
    "1620565963749101568",
    "1620566161825103873",
    "1620565275820302336",
    "1620552205131714561"
  )

laurel_tweets_jan_31 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_jan_31)

as_tibble(laurel_tweets_jan_31[["tweets"]])

as_tibble(laurel_tweets_jan_31[["users"]])

write_csv(x = laurel_tweets_jan_31[["users"]],
          file = "raw_laurel_tweets_jan_31.csv")

write_csv(x = laurel_tweets_jan_31[["tweets"]],
          file = "raw_laurel_tweets_jan_31_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1609618195585044480",
    "1609620063752970246",
    "1609707631467458560",
    "1609773785930137600",
    "1609662885780668416",
    "1609976316471947265",
    "1609685085783224322",
    "1609696350748413953",
    "1609949414776508416",
    "1609805218807676930",
    "1609981829448187905",
    "1609639162021298177",
    "1609942287542538242",
    "1609658927146467329",
    "1609658545343168515",
    "1609775655641554946",
    "1610313980790321152",
    "1609719558839078912",
    "1609684112708718592",
    "1610062400295931909",
    "1609637200500789249",
    "1609977744951070720",
    "1609930482044792833"
  )

laurel_tweets_jan_1 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_jan_1)

as_tibble(laurel_tweets_jan_1[["tweets"]])

as_tibble(laurel_tweets_jan_1[["users"]])

write_csv(x = laurel_tweets_jan_1[["users"]],
          file = "raw_laurel_tweets_jan_1.csv")

write_csv(x = laurel_tweets_jan_1[["tweets"]],
          file = "raw_laurel_tweets_jan_1_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1600881265707655168",
    "1601350468634279936",
    "1600881974754828288",
    "1600894812408143873"
  )

laurel_tweets_dec_8 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_dec_8)

as_tibble(laurel_tweets_dec_8[["tweets"]])

as_tibble(laurel_tweets_dec_8[["users"]])

write_csv(x = laurel_tweets_dec_8[["users"]],
          file = "raw_laurel_tweets_dec_8.csv")

write_csv(x = laurel_tweets_dec_8[["tweets"]],
          file = "raw_laurel_tweets_dec_8_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1596219159595323393",
    "1633139359678361603",
    "1596281062954790912",
    "1596252511530192897",
    "1596284721864245248",
    "1596656410875809792",
    "1596241763337121792"
  )

laurel_tweets_nov_25 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_nov_25)

as_tibble(laurel_tweets_nov_25[["tweets"]])

as_tibble(laurel_tweets_nov_25[["users"]])

write_csv(x = laurel_tweets_nov_25[["users"]],
          file = "raw_laurel_tweets_nov_25.csv")

write_csv(x = laurel_tweets_nov_25[["tweets"]],
          file = "raw_laurel_tweets_nov_25_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1593397410327588864",
    "1593400699567886336",
    "1593456910287503360",
    "1593414214395527168",
    "1593403843727220736"
  )

laurel_tweets_nov_17 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_nov_17)

as_tibble(laurel_tweets_nov_17[["tweets"]])

as_tibble(laurel_tweets_nov_17[["users"]])

write_csv(x = laurel_tweets_nov_17[["users"]],
          file = "raw_laurel_tweets_nov_17.csv")

write_csv(x = laurel_tweets_nov_17[["tweets"]],
          file = "raw_laurel_tweets_nov_17_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1593015623193591810",
    "1593091697395007489",
    "1593021732889452545",
    "1593281793113083910",
    "1593083661087576064",
    "1593084105193033728"
  )

laurel_tweets_nov_16 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_nov_16)

as_tibble(laurel_tweets_nov_16[["tweets"]])

as_tibble(laurel_tweets_nov_16[["users"]])

write_csv(x = laurel_tweets_nov_16[["users"]],
          file = "raw_laurel_tweets_nov_16.csv")

write_csv(x = laurel_tweets_nov_16[["tweets"]],
          file = "raw_laurel_tweets_nov_16_2.csv")

#### Laurel Collins - key word "environment", tweets from October 2022 - October 2023 #####
# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1697647149750824971",
    "1697659306873852392",
    "1697783837361840138",
    "1697661032376963655",
    "1697725891810578783",
    "1697745109608169788",
    "1697891328208900461",
    "1697673149591195855",
    "1697667997618540677",
    "1697724748564975876",
    "1697774949019320809",
    "1697750429772177769",
    "1697807093255049261",
    "1697678571907715482",
    "1697686275505992013",
    "1697769120891027850",
    "1698000490402062836",
    "1697713216019702208"
  )

laurel_tweets_sep_1 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_sep_1)

as_tibble(laurel_tweets_sep_1[["tweets"]])

as_tibble(laurel_tweets_sep_1[["users"]])

write_csv(x = laurel_tweets_sep_1[["users"]],
          file = "raw_laurel_tweets_sep_1.csv")

write_csv(x = laurel_tweets_sep_1[["tweets"]],
          file = "raw_laurel_tweets_sep_1_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1658556804224385025",
    "1658578210370486273",
    "1658557657551192064",
    "1658704353912553474",
    "1658692638684622849",
    "1658684217176489984",
    "1658693419794067459",
    "1658842921112596480",
    "1658688574265638912"
  )

laurel_tweets_may_16 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_may_16)

as_tibble(laurel_tweets_may_16[["tweets"]])

as_tibble(laurel_tweets_may_16[["users"]])

write_csv(x = laurel_tweets_may_16[["users"]],
          file = "raw_laurel_tweets_may_16.csv")

write_csv(x = laurel_tweets_may_16[["tweets"]],
          file = "raw_laurel_tweets_may_16_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1653438112461369345",
    "1653463910618656768",
    "1653468941723811855",
    "1653478009867993088"
  )

laurel_tweets_may_2 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_may_2)

as_tibble(laurel_tweets_may_2[["tweets"]])

as_tibble(laurel_tweets_may_2[["users"]])

write_csv(x = laurel_tweets_may_2[["users"]],
          file = "raw_laurel_tweets_may_2.csv")

write_csv(x = laurel_tweets_may_2[["tweets"]],
          file = "raw_laurel_tweets_may_2_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1621558740762923010",
    "1621591942982246400",
    "1621592926907408384",
    "1624094499264946182",
    "1621773852555493378",
    "1621782611638063107",
    "1621949493192310784",
    "1621617503435735040"
  )

laurel_tweets_feb_3 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_feb_3)

as_tibble(laurel_tweets_feb_3[["tweets"]])

as_tibble(laurel_tweets_feb_3[["users"]])

write_csv(x = laurel_tweets_feb_3[["users"]],
          file = "raw_laurel_tweets_feb_3.csv")

write_csv(x = laurel_tweets_feb_3[["tweets"]],
          file = "raw_laurel_tweets_feb_3_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1587577473340170240",
    "1587588692163665926"
  )

laurel_tweets_nov_1 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_nov_1)

as_tibble(laurel_tweets_nov_1[["tweets"]])

as_tibble(laurel_tweets_nov_1[["users"]])

write_csv(x = laurel_tweets_nov_1[["users"]],
          file = "raw_laurel_tweets_nov_1.csv")

write_csv(x = laurel_tweets_nov_1[["tweets"]],
          file = "raw_laurel_tweets_nov_1_2.csv")

#### Laurel Collins - key word "environmental", tweets from October 2022 - October 2023 #####
# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1658556804224385025",
    "1658578210370486273",
    "1658557657551192064",
    "1658704353912553474",
    "1658692638684622849",
    "1658684217176489984",
    "1658693419794067459",
    "1658842921112596480",
    "1658688574265638912"
  )

laurel_tweets_ma_16 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_ma_16)

as_tibble(laurel_tweets_ma_16[["tweets"]])

as_tibble(laurel_tweets_ma_16[["users"]])

write_csv(x = laurel_tweets_ma_16[["users"]],
          file = "raw_laurel_tweets_ma_16.csv")

write_csv(x = laurel_tweets_ma_16[["tweets"]],
          file = "raw_laurel_tweets_ma_16_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1649442250240450562",
    "1649961293628190720",
    "1649512016392990722",
    "1649497267949428736",
    "1649644347162521601",
    "1649552203718479874",
    "1649507160831717383",
    "1649500353115852800",
    "1649567156856844288",
    "1649589017334398976",
    "1649573326271180800"
  )

laurel_tweets_apr_21 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_apr_21)

as_tibble(laurel_tweets_apr_21[["tweets"]])

as_tibble(laurel_tweets_apr_21[["users"]])

write_csv(x = laurel_tweets_apr_21[["users"]],
          file = "raw_laurel_tweets_apr_21.csv")

write_csv(x = laurel_tweets_apr_21[["tweets"]],
          file = "raw_laurel_tweets_apr_21_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1634155882496794625",
    "1634155884329811970",
    "1634500837346271233",
    "1634688936974110720",
    "1634531685474930688",
    "1634385259742609411",
    "1634212575121661952",
    "1634245368622858241",
    "1634275350279606294",
    "1634291558966231040",
    "1634383040464953344"
  )

laurel_tweets_mar_10 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_mar_10)

as_tibble(laurel_tweets_mar_10[["tweets"]])

as_tibble(laurel_tweets_mar_10[["users"]])

write_csv(x = laurel_tweets_mar_10[["users"]],
          file = "raw_laurel_tweets_mar_10.csv")

write_csv(x = laurel_tweets_mar_10[["tweets"]],
          file = "raw_laurel_tweets_mar_10_2.csv")

# Collect tweets #
tweet_ids <-
  c(
    "https://twitter.com/Laurel_BC/status/1588593263975731200",
    "1588617405357772800",
    "1588949875764662273",
    "1588617999715807232",
    "1588610575978074112",
    "1589086726152421377",
    "1588691109802565633",
    "1588652249571815424",
    "1589034824773533697",
    "1588620087296393217",
    "1588680555310161920",
    "1588675903491825665",
    "1588677136071946241",
    "1604312471107502082",
    "1588780581936312322",
    "1588748275297157121"
  )

laurel_tweets_nov_4 <- tcn_threads(tweet_ids, token)

print(laurel_tweets_nov_4)

as_tibble(laurel_tweets_nov_4[["tweets"]])

as_tibble(laurel_tweets_nov_4[["users"]])

write_csv(x = laurel_tweets_nov_4[["users"]],
          file = "raw_laurel_tweets_nov_4.csv")

write_csv(x = laurel_tweets_nov_4[["tweets"]],
          file = "raw_laurel_tweets_nov_4_2.csv")
